@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
-// 	protoc        v4.25.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: benchmark/benchmark.proto
 
 package benchmark
@@ -12,7 +12,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 	reflect "reflect"
-	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -23,20 +23,17 @@ const (
 )
 
 type Echo struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Payload []byte `protobuf:"bytes,1,opt,name=payload,proto3" json:"payload,omitempty"`
+	state              protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Payload []byte                 `protobuf:"bytes,1,opt,name=payload"`
+	unknownFields      protoimpl.UnknownFields
+	sizeCache          protoimpl.SizeCache
 }
 
 func (x *Echo) Reset() {
 	*x = Echo{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_benchmark_benchmark_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_benchmark_benchmark_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Echo) String() string {
@@ -47,7 +44,7 @@ func (*Echo) ProtoMessage() {}
 
 func (x *Echo) ProtoReflect() protoreflect.Message {
 	mi := &file_benchmark_benchmark_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -57,34 +54,47 @@ func (x *Echo) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Echo.ProtoReflect.Descriptor instead.
-func (*Echo) Descriptor() ([]byte, []int) {
-	return file_benchmark_benchmark_proto_rawDescGZIP(), []int{0}
-}
-
 func (x *Echo) GetPayload() []byte {
 	if x != nil {
-		return x.Payload
+		return x.xxx_hidden_Payload
 	}
 	return nil
 }
 
-type TimedMsg struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
+func (x *Echo) SetPayload(v []byte) {
+	if v == nil {
+		v = []byte{}
+	}
+	x.xxx_hidden_Payload = v
+}
 
-	SendTime int64  `protobuf:"varint,1,opt,name=SendTime,proto3" json:"SendTime,omitempty"`
-	Payload  []byte `protobuf:"bytes,2,opt,name=payload,proto3" json:"payload,omitempty"`
+type Echo_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Payload []byte
+}
+
+func (b0 Echo_builder) Build() *Echo {
+	m0 := &Echo{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Payload = b.Payload
+	return m0
+}
+
+type TimedMsg struct {
+	state               protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_SendTime int64                  `protobuf:"varint,1,opt,name=SendTime"`
+	xxx_hidden_Payload  []byte                 `protobuf:"bytes,2,opt,name=payload"`
+	unknownFields       protoimpl.UnknownFields
+	sizeCache           protoimpl.SizeCache
 }
 
 func (x *TimedMsg) Reset() {
 	*x = TimedMsg{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_benchmark_benchmark_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_benchmark_benchmark_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TimedMsg) String() string {
@@ -95,7 +105,7 @@ func (*TimedMsg) ProtoMessage() {}
 
 func (x *TimedMsg) ProtoReflect() protoreflect.Message {
 	mi := &file_benchmark_benchmark_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -105,38 +115,58 @@ func (x *TimedMsg) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use TimedMsg.ProtoReflect.Descriptor instead.
-func (*TimedMsg) Descriptor() ([]byte, []int) {
-	return file_benchmark_benchmark_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *TimedMsg) GetSendTime() int64 {
 	if x != nil {
-		return x.SendTime
+		return x.xxx_hidden_SendTime
 	}
 	return 0
 }
 
 func (x *TimedMsg) GetPayload() []byte {
 	if x != nil {
-		return x.Payload
+		return x.xxx_hidden_Payload
 	}
 	return nil
 }
 
+func (x *TimedMsg) SetSendTime(v int64) {
+	x.xxx_hidden_SendTime = v
+}
+
+func (x *TimedMsg) SetPayload(v []byte) {
+	if v == nil {
+		v = []byte{}
+	}
+	x.xxx_hidden_Payload = v
+}
+
+type TimedMsg_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	SendTime int64
+	Payload  []byte
+}
+
+func (b0 TimedMsg_builder) Build() *TimedMsg {
+	m0 := &TimedMsg{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_SendTime = b.SendTime
+	x.xxx_hidden_Payload = b.Payload
+	return m0
+}
+
 type StartRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *StartRequest) Reset() {
 	*x = StartRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_benchmark_benchmark_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_benchmark_benchmark_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StartRequest) String() string {
@@ -147,7 +177,7 @@ func (*StartRequest) ProtoMessage() {}
 
 func (x *StartRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_benchmark_benchmark_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -157,24 +187,29 @@ func (x *StartRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use StartRequest.ProtoReflect.Descriptor instead.
-func (*StartRequest) Descriptor() ([]byte, []int) {
-	return file_benchmark_benchmark_proto_rawDescGZIP(), []int{2}
+type StartRequest_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 StartRequest_builder) Build() *StartRequest {
+	m0 := &StartRequest{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 type StartResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *StartResponse) Reset() {
 	*x = StartResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_benchmark_benchmark_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_benchmark_benchmark_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StartResponse) String() string {
@@ -185,7 +220,7 @@ func (*StartResponse) ProtoMessage() {}
 
 func (x *StartResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_benchmark_benchmark_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -195,24 +230,29 @@ func (x *StartResponse) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use StartResponse.ProtoReflect.Descriptor instead.
-func (*StartResponse) Descriptor() ([]byte, []int) {
-	return file_benchmark_benchmark_proto_rawDescGZIP(), []int{3}
+type StartResponse_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 StartResponse_builder) Build() *StartResponse {
+	m0 := &StartResponse{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 type StopRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *StopRequest) Reset() {
 	*x = StopRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_benchmark_benchmark_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_benchmark_benchmark_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StopRequest) String() string {
@@ -223,7 +263,7 @@ func (*StopRequest) ProtoMessage() {}
 
 func (x *StopRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_benchmark_benchmark_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -233,34 +273,38 @@ func (x *StopRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use StopRequest.ProtoReflect.Descriptor instead.
-func (*StopRequest) Descriptor() ([]byte, []int) {
-	return file_benchmark_benchmark_proto_rawDescGZIP(), []int{4}
+type StopRequest_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 StopRequest_builder) Build() *StopRequest {
+	m0 := &StopRequest{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 type Result struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Name        string        `protobuf:"bytes,1,opt,name=Name,proto3" json:"Name,omitempty"`
-	TotalOps    uint64        `protobuf:"varint,2,opt,name=TotalOps,proto3" json:"TotalOps,omitempty"`
-	TotalTime   int64         `protobuf:"varint,3,opt,name=TotalTime,proto3" json:"TotalTime,omitempty"`
-	Throughput  float64       `protobuf:"fixed64,4,opt,name=Throughput,proto3" json:"Throughput,omitempty"`
-	LatencyAvg  float64       `protobuf:"fixed64,5,opt,name=LatencyAvg,proto3" json:"LatencyAvg,omitempty"`
-	LatencyVar  float64       `protobuf:"fixed64,6,opt,name=LatencyVar,proto3" json:"LatencyVar,omitempty"`
-	AllocsPerOp uint64        `protobuf:"varint,7,opt,name=AllocsPerOp,proto3" json:"AllocsPerOp,omitempty"`
-	MemPerOp    uint64        `protobuf:"varint,8,opt,name=MemPerOp,proto3" json:"MemPerOp,omitempty"`
-	ServerStats []*MemoryStat `protobuf:"bytes,9,rep,name=ServerStats,proto3" json:"ServerStats,omitempty"`
+	state                  protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Name        string                 `protobuf:"bytes,1,opt,name=Name"`
+	xxx_hidden_TotalOps    uint64                 `protobuf:"varint,2,opt,name=TotalOps"`
+	xxx_hidden_TotalTime   int64                  `protobuf:"varint,3,opt,name=TotalTime"`
+	xxx_hidden_Throughput  float64                `protobuf:"fixed64,4,opt,name=Throughput"`
+	xxx_hidden_LatencyAvg  float64                `protobuf:"fixed64,5,opt,name=LatencyAvg"`
+	xxx_hidden_LatencyVar  float64                `protobuf:"fixed64,6,opt,name=LatencyVar"`
+	xxx_hidden_AllocsPerOp uint64                 `protobuf:"varint,7,opt,name=AllocsPerOp"`
+	xxx_hidden_MemPerOp    uint64                 `protobuf:"varint,8,opt,name=MemPerOp"`
+	xxx_hidden_ServerStats *[]*MemoryStat         `protobuf:"bytes,9,rep,name=ServerStats"`
+	unknownFields          protoimpl.UnknownFields
+	sizeCache              protoimpl.SizeCache
 }
 
 func (x *Result) Reset() {
 	*x = Result{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_benchmark_benchmark_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_benchmark_benchmark_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Result) String() string {
@@ -271,7 +315,7 @@ func (*Result) ProtoMessage() {}
 
 func (x *Result) ProtoReflect() protoreflect.Message {
 	mi := &file_benchmark_benchmark_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -281,90 +325,150 @@ func (x *Result) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Result.ProtoReflect.Descriptor instead.
-func (*Result) Descriptor() ([]byte, []int) {
-	return file_benchmark_benchmark_proto_rawDescGZIP(), []int{5}
-}
-
 func (x *Result) GetName() string {
 	if x != nil {
-		return x.Name
+		return x.xxx_hidden_Name
 	}
 	return ""
 }
 
 func (x *Result) GetTotalOps() uint64 {
 	if x != nil {
-		return x.TotalOps
+		return x.xxx_hidden_TotalOps
 	}
 	return 0
 }
 
 func (x *Result) GetTotalTime() int64 {
 	if x != nil {
-		return x.TotalTime
+		return x.xxx_hidden_TotalTime
 	}
 	return 0
 }
 
 func (x *Result) GetThroughput() float64 {
 	if x != nil {
-		return x.Throughput
+		return x.xxx_hidden_Throughput
 	}
 	return 0
 }
 
 func (x *Result) GetLatencyAvg() float64 {
 	if x != nil {
-		return x.LatencyAvg
+		return x.xxx_hidden_LatencyAvg
 	}
 	return 0
 }
 
 func (x *Result) GetLatencyVar() float64 {
 	if x != nil {
-		return x.LatencyVar
+		return x.xxx_hidden_LatencyVar
 	}
 	return 0
 }
 
 func (x *Result) GetAllocsPerOp() uint64 {
 	if x != nil {
-		return x.AllocsPerOp
+		return x.xxx_hidden_AllocsPerOp
 	}
 	return 0
 }
 
 func (x *Result) GetMemPerOp() uint64 {
 	if x != nil {
-		return x.MemPerOp
+		return x.xxx_hidden_MemPerOp
 	}
 	return 0
 }
 
 func (x *Result) GetServerStats() []*MemoryStat {
 	if x != nil {
-		return x.ServerStats
+		if x.xxx_hidden_ServerStats != nil {
+			return *x.xxx_hidden_ServerStats
+		}
 	}
 	return nil
 }
 
-type MemoryStat struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
+func (x *Result) SetName(v string) {
+	x.xxx_hidden_Name = v
+}
 
-	Allocs uint64 `protobuf:"varint,1,opt,name=Allocs,proto3" json:"Allocs,omitempty"`
-	Memory uint64 `protobuf:"varint,2,opt,name=Memory,proto3" json:"Memory,omitempty"`
+func (x *Result) SetTotalOps(v uint64) {
+	x.xxx_hidden_TotalOps = v
+}
+
+func (x *Result) SetTotalTime(v int64) {
+	x.xxx_hidden_TotalTime = v
+}
+
+func (x *Result) SetThroughput(v float64) {
+	x.xxx_hidden_Throughput = v
+}
+
+func (x *Result) SetLatencyAvg(v float64) {
+	x.xxx_hidden_LatencyAvg = v
+}
+
+func (x *Result) SetLatencyVar(v float64) {
+	x.xxx_hidden_LatencyVar = v
+}
+
+func (x *Result) SetAllocsPerOp(v uint64) {
+	x.xxx_hidden_AllocsPerOp = v
+}
+
+func (x *Result) SetMemPerOp(v uint64) {
+	x.xxx_hidden_MemPerOp = v
+}
+
+func (x *Result) SetServerStats(v []*MemoryStat) {
+	x.xxx_hidden_ServerStats = &v
+}
+
+type Result_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Name        string
+	TotalOps    uint64
+	TotalTime   int64
+	Throughput  float64
+	LatencyAvg  float64
+	LatencyVar  float64
+	AllocsPerOp uint64
+	MemPerOp    uint64
+	ServerStats []*MemoryStat
+}
+
+func (b0 Result_builder) Build() *Result {
+	m0 := &Result{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Name = b.Name
+	x.xxx_hidden_TotalOps = b.TotalOps
+	x.xxx_hidden_TotalTime = b.TotalTime
+	x.xxx_hidden_Throughput = b.Throughput
+	x.xxx_hidden_LatencyAvg = b.LatencyAvg
+	x.xxx_hidden_LatencyVar = b.LatencyVar
+	x.xxx_hidden_AllocsPerOp = b.AllocsPerOp
+	x.xxx_hidden_MemPerOp = b.MemPerOp
+	x.xxx_hidden_ServerStats = &b.ServerStats
+	return m0
+}
+
+type MemoryStat struct {
+	state             protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Allocs uint64                 `protobuf:"varint,1,opt,name=Allocs"`
+	xxx_hidden_Memory uint64                 `protobuf:"varint,2,opt,name=Memory"`
+	unknownFields     protoimpl.UnknownFields
+	sizeCache         protoimpl.SizeCache
 }
 
 func (x *MemoryStat) Reset() {
 	*x = MemoryStat{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_benchmark_benchmark_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_benchmark_benchmark_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MemoryStat) String() string {
@@ -375,7 +479,7 @@ func (*MemoryStat) ProtoMessage() {}
 
 func (x *MemoryStat) ProtoReflect() protoreflect.Message {
 	mi := &file_benchmark_benchmark_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -385,40 +489,56 @@ func (x *MemoryStat) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use MemoryStat.ProtoReflect.Descriptor instead.
-func (*MemoryStat) Descriptor() ([]byte, []int) {
-	return file_benchmark_benchmark_proto_rawDescGZIP(), []int{6}
-}
-
 func (x *MemoryStat) GetAllocs() uint64 {
 	if x != nil {
-		return x.Allocs
+		return x.xxx_hidden_Allocs
 	}
 	return 0
 }
 
 func (x *MemoryStat) GetMemory() uint64 {
 	if x != nil {
-		return x.Memory
+		return x.xxx_hidden_Memory
 	}
 	return 0
 }
 
-type MemoryStatList struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
+func (x *MemoryStat) SetAllocs(v uint64) {
+	x.xxx_hidden_Allocs = v
+}
 
-	MemoryStats []*MemoryStat `protobuf:"bytes,1,rep,name=MemoryStats,proto3" json:"MemoryStats,omitempty"`
+func (x *MemoryStat) SetMemory(v uint64) {
+	x.xxx_hidden_Memory = v
+}
+
+type MemoryStat_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Allocs uint64
+	Memory uint64
+}
+
+func (b0 MemoryStat_builder) Build() *MemoryStat {
+	m0 := &MemoryStat{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Allocs = b.Allocs
+	x.xxx_hidden_Memory = b.Memory
+	return m0
+}
+
+type MemoryStatList struct {
+	state                  protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_MemoryStats *[]*MemoryStat         `protobuf:"bytes,1,rep,name=MemoryStats"`
+	unknownFields          protoimpl.UnknownFields
+	sizeCache              protoimpl.SizeCache
 }
 
 func (x *MemoryStatList) Reset() {
 	*x = MemoryStatList{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_benchmark_benchmark_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_benchmark_benchmark_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MemoryStatList) String() string {
@@ -429,7 +549,7 @@ func (*MemoryStatList) ProtoMessage() {}
 
 func (x *MemoryStatList) ProtoReflect() protoreflect.Message {
 	mi := &file_benchmark_benchmark_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -439,117 +559,82 @@ func (x *MemoryStatList) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use MemoryStatList.ProtoReflect.Descriptor instead.
-func (*MemoryStatList) Descriptor() ([]byte, []int) {
-	return file_benchmark_benchmark_proto_rawDescGZIP(), []int{7}
-}
-
 func (x *MemoryStatList) GetMemoryStats() []*MemoryStat {
 	if x != nil {
-		return x.MemoryStats
+		if x.xxx_hidden_MemoryStats != nil {
+			return *x.xxx_hidden_MemoryStats
+		}
 	}
 	return nil
 }
 
+func (x *MemoryStatList) SetMemoryStats(v []*MemoryStat) {
+	x.xxx_hidden_MemoryStats = &v
+}
+
+type MemoryStatList_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	MemoryStats []*MemoryStat
+}
+
+func (b0 MemoryStatList_builder) Build() *MemoryStatList {
+	m0 := &MemoryStatList{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_MemoryStats = &b.MemoryStats
+	return m0
+}
+
 var File_benchmark_benchmark_proto protoreflect.FileDescriptor
 
-var file_benchmark_benchmark_proto_rawDesc = []byte{
-	0x0a, 0x19, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2f, 0x62, 0x65, 0x6e, 0x63,
-	0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x09, 0x62, 0x65, 0x6e,
-	0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x1a, 0x0c, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x22, 0x20, 0x0a, 0x04, 0x45, 0x63, 0x68, 0x6f, 0x12, 0x18, 0x0a, 0x07, 0x70, 0x61, 0x79,
-	0x6c, 0x6f, 0x61, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x07, 0x70, 0x61, 0x79, 0x6c,
-	0x6f, 0x61, 0x64, 0x22, 0x40, 0x0a, 0x08, 0x54, 0x69, 0x6d, 0x65, 0x64, 0x4d, 0x73, 0x67, 0x12,
-	0x1a, 0x0a, 0x08, 0x53, 0x65, 0x6e, 0x64, 0x54, 0x69, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x03, 0x52, 0x08, 0x53, 0x65, 0x6e, 0x64, 0x54, 0x69, 0x6d, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x70,
-	0x61, 0x79, 0x6c, 0x6f, 0x61, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x07, 0x70, 0x61,
-	0x79, 0x6c, 0x6f, 0x61, 0x64, 0x22, 0x0e, 0x0a, 0x0c, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x0f, 0x0a, 0x0d, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x0d, 0x0a, 0x0b, 0x53, 0x74, 0x6f, 0x70, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0xad, 0x02, 0x0a, 0x06, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74,
-	0x12, 0x12, 0x0a, 0x04, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04,
-	0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x54, 0x6f, 0x74, 0x61, 0x6c, 0x4f, 0x70, 0x73,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x08, 0x54, 0x6f, 0x74, 0x61, 0x6c, 0x4f, 0x70, 0x73,
-	0x12, 0x1c, 0x0a, 0x09, 0x54, 0x6f, 0x74, 0x61, 0x6c, 0x54, 0x69, 0x6d, 0x65, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x03, 0x52, 0x09, 0x54, 0x6f, 0x74, 0x61, 0x6c, 0x54, 0x69, 0x6d, 0x65, 0x12, 0x1e,
-	0x0a, 0x0a, 0x54, 0x68, 0x72, 0x6f, 0x75, 0x67, 0x68, 0x70, 0x75, 0x74, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x01, 0x52, 0x0a, 0x54, 0x68, 0x72, 0x6f, 0x75, 0x67, 0x68, 0x70, 0x75, 0x74, 0x12, 0x1e,
-	0x0a, 0x0a, 0x4c, 0x61, 0x74, 0x65, 0x6e, 0x63, 0x79, 0x41, 0x76, 0x67, 0x18, 0x05, 0x20, 0x01,
-	0x28, 0x01, 0x52, 0x0a, 0x4c, 0x61, 0x74, 0x65, 0x6e, 0x63, 0x79, 0x41, 0x76, 0x67, 0x12, 0x1e,
-	0x0a, 0x0a, 0x4c, 0x61, 0x74, 0x65, 0x6e, 0x63, 0x79, 0x56, 0x61, 0x72, 0x18, 0x06, 0x20, 0x01,
-	0x28, 0x01, 0x52, 0x0a, 0x4c, 0x61, 0x74, 0x65, 0x6e, 0x63, 0x79, 0x56, 0x61, 0x72, 0x12, 0x20,
-	0x0a, 0x0b, 0x41, 0x6c, 0x6c, 0x6f, 0x63, 0x73, 0x50, 0x65, 0x72, 0x4f, 0x70, 0x18, 0x07, 0x20,
-	0x01, 0x28, 0x04, 0x52, 0x0b, 0x41, 0x6c, 0x6c, 0x6f, 0x63, 0x73, 0x50, 0x65, 0x72, 0x4f, 0x70,
-	0x12, 0x1a, 0x0a, 0x08, 0x4d, 0x65, 0x6d, 0x50, 0x65, 0x72, 0x4f, 0x70, 0x18, 0x08, 0x20, 0x01,
-	0x28, 0x04, 0x52, 0x08, 0x4d, 0x65, 0x6d, 0x50, 0x65, 0x72, 0x4f, 0x70, 0x12, 0x37, 0x0a, 0x0b,
-	0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x53, 0x74, 0x61, 0x74, 0x73, 0x18, 0x09, 0x20, 0x03, 0x28,
-	0x0b, 0x32, 0x15, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x4d, 0x65,
-	0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61, 0x74, 0x52, 0x0b, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72,
-	0x53, 0x74, 0x61, 0x74, 0x73, 0x22, 0x3c, 0x0a, 0x0a, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53,
-	0x74, 0x61, 0x74, 0x12, 0x16, 0x0a, 0x06, 0x41, 0x6c, 0x6c, 0x6f, 0x63, 0x73, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x04, 0x52, 0x06, 0x41, 0x6c, 0x6c, 0x6f, 0x63, 0x73, 0x12, 0x16, 0x0a, 0x06, 0x4d,
-	0x65, 0x6d, 0x6f, 0x72, 0x79, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x06, 0x4d, 0x65, 0x6d,
-	0x6f, 0x72, 0x79, 0x22, 0x49, 0x0a, 0x0e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61,
-	0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x37, 0x0a, 0x0b, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53,
-	0x74, 0x61, 0x74, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x62, 0x65, 0x6e,
-	0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61,
-	0x74, 0x52, 0x0b, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61, 0x74, 0x73, 0x32, 0xb2,
-	0x04, 0x0a, 0x09, 0x42, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x12, 0x4f, 0x0a, 0x14,
-	0x53, 0x74, 0x61, 0x72, 0x74, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x42, 0x65, 0x6e, 0x63, 0x68,
-	0x6d, 0x61, 0x72, 0x6b, 0x12, 0x17, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b,
-	0x2e, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x18, 0x2e,
-	0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x04, 0xa0, 0xb5, 0x18, 0x01, 0x12, 0x46, 0x0a,
-	0x13, 0x53, 0x74, 0x6f, 0x70, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x42, 0x65, 0x6e, 0x63, 0x68,
-	0x6d, 0x61, 0x72, 0x6b, 0x12, 0x16, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b,
-	0x2e, 0x53, 0x74, 0x6f, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x11, 0x2e, 0x62,
-	0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x22,
-	0x04, 0xa0, 0xb5, 0x18, 0x01, 0x12, 0x49, 0x0a, 0x0e, 0x53, 0x74, 0x61, 0x72, 0x74, 0x42, 0x65,
-	0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x12, 0x17, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d,
-	0x61, 0x72, 0x6b, 0x2e, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
-	0x1a, 0x18, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x53, 0x74, 0x61,
-	0x72, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x04, 0xa0, 0xb5, 0x18, 0x01,
-	0x12, 0x56, 0x0a, 0x0d, 0x53, 0x74, 0x6f, 0x70, 0x42, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72,
-	0x6b, 0x12, 0x16, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x53, 0x74,
-	0x6f, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x15, 0x2e, 0x62, 0x65, 0x6e, 0x63,
-	0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61, 0x74,
-	0x22, 0x16, 0xa0, 0xb5, 0x18, 0x01, 0xf2, 0xb6, 0x18, 0x0e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79,
-	0x53, 0x74, 0x61, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x34, 0x0a, 0x0a, 0x51, 0x75, 0x6f, 0x72,
-	0x75, 0x6d, 0x43, 0x61, 0x6c, 0x6c, 0x12, 0x0f, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61,
-	0x72, 0x6b, 0x2e, 0x45, 0x63, 0x68, 0x6f, 0x1a, 0x0f, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d,
-	0x61, 0x72, 0x6b, 0x2e, 0x45, 0x63, 0x68, 0x6f, 0x22, 0x04, 0xa0, 0xb5, 0x18, 0x01, 0x12, 0x3d,
-	0x0a, 0x0f, 0x41, 0x73, 0x79, 0x6e, 0x63, 0x51, 0x75, 0x6f, 0x72, 0x75, 0x6d, 0x43, 0x61, 0x6c,
-	0x6c, 0x12, 0x0f, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x45, 0x63,
-	0x68, 0x6f, 0x1a, 0x0f, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x45,
-	0x63, 0x68, 0x6f, 0x22, 0x08, 0xa0, 0xb5, 0x18, 0x01, 0xd0, 0xb5, 0x18, 0x01, 0x12, 0x34, 0x0a,
-	0x0a, 0x53, 0x6c, 0x6f, 0x77, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x12, 0x0f, 0x2e, 0x62, 0x65,
-	0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x45, 0x63, 0x68, 0x6f, 0x1a, 0x0f, 0x2e, 0x62,
-	0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x45, 0x63, 0x68, 0x6f, 0x22, 0x04, 0xa0,
-	0xb5, 0x18, 0x01, 0x12, 0x3e, 0x0a, 0x09, 0x4d, 0x75, 0x6c, 0x74, 0x69, 0x63, 0x61, 0x73, 0x74,
-	0x12, 0x13, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x54, 0x69, 0x6d,
-	0x65, 0x64, 0x4d, 0x73, 0x67, 0x1a, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x04, 0x98,
-	0xb5, 0x18, 0x01, 0x42, 0x23, 0x5a, 0x21, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f,
-	0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x62,
-	0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
-
-var (
-	file_benchmark_benchmark_proto_rawDescOnce sync.Once
-	file_benchmark_benchmark_proto_rawDescData = file_benchmark_benchmark_proto_rawDesc
-)
-
-func file_benchmark_benchmark_proto_rawDescGZIP() []byte {
-	file_benchmark_benchmark_proto_rawDescOnce.Do(func() {
-		file_benchmark_benchmark_proto_rawDescData = protoimpl.X.CompressGZIP(file_benchmark_benchmark_proto_rawDescData)
-	})
-	return file_benchmark_benchmark_proto_rawDescData
-}
+const file_benchmark_benchmark_proto_rawDesc = "" +
+	"\n" +
+	"\x19benchmark/benchmark.proto\x12\tbenchmark\x1a\x1bgoogle/protobuf/empty.proto\x1a\fgorums.proto\" \n" +
+	"\x04Echo\x12\x18\n" +
+	"\apayload\x18\x01 \x01(\fR\apayload\"@\n" +
+	"\bTimedMsg\x12\x1a\n" +
+	"\bSendTime\x18\x01 \x01(\x03R\bSendTime\x12\x18\n" +
+	"\apayload\x18\x02 \x01(\fR\apayload\"\x0e\n" +
+	"\fStartRequest\"\x0f\n" +
+	"\rStartResponse\"\r\n" +
+	"\vStopRequest\"\xad\x02\n" +
+	"\x06Result\x12\x12\n" +
+	"\x04Name\x18\x01 \x01(\tR\x04Name\x12\x1a\n" +
+	"\bTotalOps\x18\x02 \x01(\x04R\bTotalOps\x12\x1c\n" +
+	"\tTotalTime\x18\x03 \x01(\x03R\tTotalTime\x12\x1e\n" +
+	"\n" +
+	"Throughput\x18\x04 \x01(\x01R\n" +
+	"Throughput\x12\x1e\n" +
+	"\n" +
+	"LatencyAvg\x18\x05 \x01(\x01R\n" +
+	"LatencyAvg\x12\x1e\n" +
+	"\n" +
+	"LatencyVar\x18\x06 \x01(\x01R\n" +
+	"LatencyVar\x12 \n" +
+	"\vAllocsPerOp\x18\a \x01(\x04R\vAllocsPerOp\x12\x1a\n" +
+	"\bMemPerOp\x18\b \x01(\x04R\bMemPerOp\x127\n" +
+	"\vServerStats\x18\t \x03(\v2\x15.benchmark.MemoryStatR\vServerStats\"<\n" +
+	"\n" +
+	"MemoryStat\x12\x16\n" +
+	"\x06Allocs\x18\x01 \x01(\x04R\x06Allocs\x12\x16\n" +
+	"\x06Memory\x18\x02 \x01(\x04R\x06Memory\"I\n" +
+	"\x0eMemoryStatList\x127\n" +
+	"\vMemoryStats\x18\x01 \x03(\v2\x15.benchmark.MemoryStatR\vMemoryStats2\xb2\x04\n" +
+	"\tBenchmark\x12O\n" +
+	"\x14StartServerBenchmark\x12\x17.benchmark.StartRequest\x1a\x18.benchmark.StartResponse\"\x04\xa0\xb5\x18\x01\x12F\n" +
+	"\x13StopServerBenchmark\x12\x16.benchmark.StopRequest\x1a\x11.benchmark.Result\"\x04\xa0\xb5\x18\x01\x12I\n" +
+	"\x0eStartBenchmark\x12\x17.benchmark.StartRequest\x1a\x18.benchmark.StartResponse\"\x04\xa0\xb5\x18\x01\x12V\n" +
+	"\rStopBenchmark\x12\x16.benchmark.StopRequest\x1a\x15.benchmark.MemoryStat\"\x16\xa0\xb5\x18\x01\xf2\xb6\x18\x0eMemoryStatList\x124\n" +
+	"\n" +
+	"QuorumCall\x12\x0f.benchmark.Echo\x1a\x0f.benchmark.Echo\"\x04\xa0\xb5\x18\x01\x12=\n" +
+	"\x0fAsyncQuorumCall\x12\x0f.benchmark.Echo\x1a\x0f.benchmark.Echo\"\b\xa0\xb5\x18\x01е\x18\x01\x124\n" +
+	"\n" +
+	"SlowServer\x12\x0f.benchmark.Echo\x1a\x0f.benchmark.Echo\"\x04\xa0\xb5\x18\x01\x12>\n" +
+	"\tMulticast\x12\x13.benchmark.TimedMsg\x1a\x16.google.protobuf.Empty\"\x04\x98\xb5\x18\x01B(Z!github.com/relab/gorums/benchmark\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_benchmark_benchmark_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
-var file_benchmark_benchmark_proto_goTypes = []interface{}{
+var file_benchmark_benchmark_proto_goTypes = []any{
 	(*Echo)(nil),           // 0: benchmark.Echo
 	(*TimedMsg)(nil),       // 1: benchmark.TimedMsg
 	(*StartRequest)(nil),   // 2: benchmark.StartRequest
@@ -591,109 +676,11 @@ func file_benchmark_benchmark_proto_init() {
 	if File_benchmark_benchmark_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_benchmark_benchmark_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Echo); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_benchmark_benchmark_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TimedMsg); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_benchmark_benchmark_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StartRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_benchmark_benchmark_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StartResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_benchmark_benchmark_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StopRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_benchmark_benchmark_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Result); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_benchmark_benchmark_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MemoryStat); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_benchmark_benchmark_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MemoryStatList); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_benchmark_benchmark_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_benchmark_benchmark_proto_rawDesc), len(file_benchmark_benchmark_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   8,
 			NumExtensions: 0,
@@ -704,7 +691,6 @@ func file_benchmark_benchmark_proto_init() {
 		MessageInfos:      file_benchmark_benchmark_proto_msgTypes,
 	}.Build()
 	File_benchmark_benchmark_proto = out.File
-	file_benchmark_benchmark_proto_rawDesc = nil
 	file_benchmark_benchmark_proto_goTypes = nil
 	file_benchmark_benchmark_proto_depIdxs = nil
 }
