@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
-// 	protoc        v4.25.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: storage/proto/storage.proto
 
 package proto
@@ -13,7 +13,7 @@ import (
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 	reflect "reflect"
-	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -24,20 +24,17 @@ const (
 )
 
 type ReadRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Key string `protobuf:"bytes,1,opt,name=Key,proto3" json:"Key,omitempty"`
+	state          protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Key string                 `protobuf:"bytes,1,opt,name=Key"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *ReadRequest) Reset() {
 	*x = ReadRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_storage_proto_storage_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_storage_proto_storage_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ReadRequest) String() string {
@@ -48,7 +45,7 @@ func (*ReadRequest) ProtoMessage() {}
 
 func (x *ReadRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_storage_proto_storage_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -58,35 +55,45 @@ func (x *ReadRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use ReadRequest.ProtoReflect.Descriptor instead.
-func (*ReadRequest) Descriptor() ([]byte, []int) {
-	return file_storage_proto_storage_proto_rawDescGZIP(), []int{0}
-}
-
 func (x *ReadRequest) GetKey() string {
 	if x != nil {
-		return x.Key
+		return x.xxx_hidden_Key
 	}
 	return ""
 }
 
-type ReadResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
+func (x *ReadRequest) SetKey(v string) {
+	x.xxx_hidden_Key = v
+}
 
-	OK    bool                   `protobuf:"varint,1,opt,name=OK,proto3" json:"OK,omitempty"`
-	Value string                 `protobuf:"bytes,2,opt,name=Value,proto3" json:"Value,omitempty"`
-	Time  *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=Time,proto3" json:"Time,omitempty"`
+type ReadRequest_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Key string
+}
+
+func (b0 ReadRequest_builder) Build() *ReadRequest {
+	m0 := &ReadRequest{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Key = b.Key
+	return m0
+}
+
+type ReadResponse struct {
+	state            protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_OK    bool                   `protobuf:"varint,1,opt,name=OK"`
+	xxx_hidden_Value string                 `protobuf:"bytes,2,opt,name=Value"`
+	xxx_hidden_Time  *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=Time"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *ReadResponse) Reset() {
 	*x = ReadResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_storage_proto_storage_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_storage_proto_storage_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ReadResponse) String() string {
@@ -97,7 +104,7 @@ func (*ReadResponse) ProtoMessage() {}
 
 func (x *ReadResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_storage_proto_storage_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -107,49 +114,82 @@ func (x *ReadResponse) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use ReadResponse.ProtoReflect.Descriptor instead.
-func (*ReadResponse) Descriptor() ([]byte, []int) {
-	return file_storage_proto_storage_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *ReadResponse) GetOK() bool {
 	if x != nil {
-		return x.OK
+		return x.xxx_hidden_OK
 	}
 	return false
 }
 
 func (x *ReadResponse) GetValue() string {
 	if x != nil {
-		return x.Value
+		return x.xxx_hidden_Value
 	}
 	return ""
 }
 
 func (x *ReadResponse) GetTime() *timestamppb.Timestamp {
 	if x != nil {
-		return x.Time
+		return x.xxx_hidden_Time
 	}
 	return nil
 }
 
-type WriteRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
+func (x *ReadResponse) SetOK(v bool) {
+	x.xxx_hidden_OK = v
+}
 
-	Key   string                 `protobuf:"bytes,1,opt,name=Key,proto3" json:"Key,omitempty"`
-	Value string                 `protobuf:"bytes,2,opt,name=Value,proto3" json:"Value,omitempty"`
-	Time  *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=Time,proto3" json:"Time,omitempty"`
+func (x *ReadResponse) SetValue(v string) {
+	x.xxx_hidden_Value = v
+}
+
+func (x *ReadResponse) SetTime(v *timestamppb.Timestamp) {
+	x.xxx_hidden_Time = v
+}
+
+func (x *ReadResponse) HasTime() bool {
+	if x == nil {
+		return false
+	}
+	return x.xxx_hidden_Time != nil
+}
+
+func (x *ReadResponse) ClearTime() {
+	x.xxx_hidden_Time = nil
+}
+
+type ReadResponse_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	OK    bool
+	Value string
+	Time  *timestamppb.Timestamp
+}
+
+func (b0 ReadResponse_builder) Build() *ReadResponse {
+	m0 := &ReadResponse{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_OK = b.OK
+	x.xxx_hidden_Value = b.Value
+	x.xxx_hidden_Time = b.Time
+	return m0
+}
+
+type WriteRequest struct {
+	state            protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Key   string                 `protobuf:"bytes,1,opt,name=Key"`
+	xxx_hidden_Value string                 `protobuf:"bytes,2,opt,name=Value"`
+	xxx_hidden_Time  *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=Time"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *WriteRequest) Reset() {
 	*x = WriteRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_storage_proto_storage_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_storage_proto_storage_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *WriteRequest) String() string {
@@ -160,7 +200,7 @@ func (*WriteRequest) ProtoMessage() {}
 
 func (x *WriteRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_storage_proto_storage_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -170,47 +210,80 @@ func (x *WriteRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use WriteRequest.ProtoReflect.Descriptor instead.
-func (*WriteRequest) Descriptor() ([]byte, []int) {
-	return file_storage_proto_storage_proto_rawDescGZIP(), []int{2}
-}
-
 func (x *WriteRequest) GetKey() string {
 	if x != nil {
-		return x.Key
+		return x.xxx_hidden_Key
 	}
 	return ""
 }
 
 func (x *WriteRequest) GetValue() string {
 	if x != nil {
-		return x.Value
+		return x.xxx_hidden_Value
 	}
 	return ""
 }
 
 func (x *WriteRequest) GetTime() *timestamppb.Timestamp {
 	if x != nil {
-		return x.Time
+		return x.xxx_hidden_Time
 	}
 	return nil
 }
 
-type WriteResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
+func (x *WriteRequest) SetKey(v string) {
+	x.xxx_hidden_Key = v
+}
 
-	New bool `protobuf:"varint,1,opt,name=New,proto3" json:"New,omitempty"`
+func (x *WriteRequest) SetValue(v string) {
+	x.xxx_hidden_Value = v
+}
+
+func (x *WriteRequest) SetTime(v *timestamppb.Timestamp) {
+	x.xxx_hidden_Time = v
+}
+
+func (x *WriteRequest) HasTime() bool {
+	if x == nil {
+		return false
+	}
+	return x.xxx_hidden_Time != nil
+}
+
+func (x *WriteRequest) ClearTime() {
+	x.xxx_hidden_Time = nil
+}
+
+type WriteRequest_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Key   string
+	Value string
+	Time  *timestamppb.Timestamp
+}
+
+func (b0 WriteRequest_builder) Build() *WriteRequest {
+	m0 := &WriteRequest{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Key = b.Key
+	x.xxx_hidden_Value = b.Value
+	x.xxx_hidden_Time = b.Time
+	return m0
+}
+
+type WriteResponse struct {
+	state          protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_New bool                   `protobuf:"varint,1,opt,name=New"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *WriteResponse) Reset() {
 	*x = WriteResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_storage_proto_storage_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_storage_proto_storage_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *WriteResponse) String() string {
@@ -221,7 +294,7 @@ func (*WriteResponse) ProtoMessage() {}
 
 func (x *WriteResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_storage_proto_storage_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -231,86 +304,57 @@ func (x *WriteResponse) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use WriteResponse.ProtoReflect.Descriptor instead.
-func (*WriteResponse) Descriptor() ([]byte, []int) {
-	return file_storage_proto_storage_proto_rawDescGZIP(), []int{3}
-}
-
 func (x *WriteResponse) GetNew() bool {
 	if x != nil {
-		return x.New
+		return x.xxx_hidden_New
 	}
 	return false
 }
 
+func (x *WriteResponse) SetNew(v bool) {
+	x.xxx_hidden_New = v
+}
+
+type WriteResponse_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	New bool
+}
+
+func (b0 WriteResponse_builder) Build() *WriteResponse {
+	m0 := &WriteResponse{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_New = b.New
+	return m0
+}
+
 var File_storage_proto_storage_proto protoreflect.FileDescriptor
 
-var file_storage_proto_storage_proto_rawDesc = []byte{
-	0x0a, 0x1b, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f,
-	0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x07, 0x73,
-	0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x1a, 0x0c, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x22, 0x1f, 0x0a, 0x0b, 0x52, 0x65, 0x61, 0x64, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x12, 0x10, 0x0a, 0x03, 0x4b, 0x65, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03,
-	0x4b, 0x65, 0x79, 0x22, 0x64, 0x0a, 0x0c, 0x52, 0x65, 0x61, 0x64, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x12, 0x0e, 0x0a, 0x02, 0x4f, 0x4b, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52,
-	0x02, 0x4f, 0x4b, 0x12, 0x14, 0x0a, 0x05, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x05, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x12, 0x2e, 0x0a, 0x04, 0x54, 0x69, 0x6d,
-	0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74,
-	0x61, 0x6d, 0x70, 0x52, 0x04, 0x54, 0x69, 0x6d, 0x65, 0x22, 0x66, 0x0a, 0x0c, 0x57, 0x72, 0x69,
-	0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x4b, 0x65, 0x79,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x4b, 0x65, 0x79, 0x12, 0x14, 0x0a, 0x05, 0x56,
-	0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x56, 0x61, 0x6c, 0x75,
-	0x65, 0x12, 0x2e, 0x0a, 0x04, 0x54, 0x69, 0x6d, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x04, 0x54, 0x69, 0x6d,
-	0x65, 0x22, 0x21, 0x0a, 0x0d, 0x57, 0x72, 0x69, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
-	0x73, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x4e, 0x65, 0x77, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52,
-	0x03, 0x4e, 0x65, 0x77, 0x32, 0xc4, 0x02, 0x0a, 0x07, 0x53, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65,
-	0x12, 0x38, 0x0a, 0x07, 0x52, 0x65, 0x61, 0x64, 0x52, 0x50, 0x43, 0x12, 0x14, 0x2e, 0x73, 0x74,
-	0x6f, 0x72, 0x61, 0x67, 0x65, 0x2e, 0x52, 0x65, 0x61, 0x64, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x1a, 0x15, 0x2e, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x2e, 0x52, 0x65, 0x61, 0x64,
-	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x3b, 0x0a, 0x08, 0x57, 0x72,
-	0x69, 0x74, 0x65, 0x52, 0x50, 0x43, 0x12, 0x15, 0x2e, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65,
-	0x2e, 0x57, 0x72, 0x69, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e,
-	0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x2e, 0x57, 0x72, 0x69, 0x74, 0x65, 0x52, 0x65, 0x73,
-	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x3b, 0x0a, 0x06, 0x52, 0x65, 0x61, 0x64, 0x51,
-	0x43, 0x12, 0x14, 0x2e, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x2e, 0x52, 0x65, 0x61, 0x64,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x15, 0x2e, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67,
-	0x65, 0x2e, 0x52, 0x65, 0x61, 0x64, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x04,
-	0xa0, 0xb5, 0x18, 0x01, 0x12, 0x3e, 0x0a, 0x07, 0x57, 0x72, 0x69, 0x74, 0x65, 0x51, 0x43, 0x12,
-	0x15, 0x2e, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x2e, 0x57, 0x72, 0x69, 0x74, 0x65, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65,
-	0x2e, 0x57, 0x72, 0x69, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x04,
-	0xa0, 0xb5, 0x18, 0x01, 0x12, 0x45, 0x0a, 0x0e, 0x57, 0x72, 0x69, 0x74, 0x65, 0x4d, 0x75, 0x6c,
-	0x74, 0x69, 0x63, 0x61, 0x73, 0x74, 0x12, 0x15, 0x2e, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65,
-	0x2e, 0x57, 0x72, 0x69, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e,
-	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
-	0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x04, 0x98, 0xb5, 0x18, 0x01, 0x42, 0x30, 0x5a, 0x2e, 0x67,
-	0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f,
-	0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x65, 0x78, 0x61, 0x6d, 0x70, 0x6c, 0x65, 0x73, 0x2f,
-	0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
-
-var (
-	file_storage_proto_storage_proto_rawDescOnce sync.Once
-	file_storage_proto_storage_proto_rawDescData = file_storage_proto_storage_proto_rawDesc
-)
-
-func file_storage_proto_storage_proto_rawDescGZIP() []byte {
-	file_storage_proto_storage_proto_rawDescOnce.Do(func() {
-		file_storage_proto_storage_proto_rawDescData = protoimpl.X.CompressGZIP(file_storage_proto_storage_proto_rawDescData)
-	})
-	return file_storage_proto_storage_proto_rawDescData
-}
+const file_storage_proto_storage_proto_rawDesc = "" +
+	"\n" +
+	"\x1bstorage/proto/storage.proto\x12\astorage\x1a\fgorums.proto\x1a\x1fgoogle/protobuf/timestamp.proto\x1a\x1bgoogle/protobuf/empty.proto\"\x1f\n" +
+	"\vReadRequest\x12\x10\n" +
+	"\x03Key\x18\x01 \x01(\tR\x03Key\"d\n" +
+	"\fReadResponse\x12\x0e\n" +
+	"\x02OK\x18\x01 \x01(\bR\x02OK\x12\x14\n" +
+	"\x05Value\x18\x02 \x01(\tR\x05Value\x12.\n" +
+	"\x04Time\x18\x03 \x01(\v2\x1a.google.protobuf.TimestampR\x04Time\"f\n" +
+	"\fWriteRequest\x12\x10\n" +
+	"\x03Key\x18\x01 \x01(\tR\x03Key\x12\x14\n" +
+	"\x05Value\x18\x02 \x01(\tR\x05Value\x12.\n" +
+	"\x04Time\x18\x03 \x01(\v2\x1a.google.protobuf.TimestampR\x04Time\"!\n" +
+	"\rWriteResponse\x12\x10\n" +
+	"\x03New\x18\x01 \x01(\bR\x03New2\xc4\x02\n" +
+	"\aStorage\x128\n" +
+	"\aReadRPC\x12\x14.storage.ReadRequest\x1a\x15.storage.ReadResponse\"\x00\x12;\n" +
+	"\bWriteRPC\x12\x15.storage.WriteRequest\x1a\x16.storage.WriteResponse\"\x00\x12;\n" +
+	"\x06ReadQC\x12\x14.storage.ReadRequest\x1a\x15.storage.ReadResponse\"\x04\xa0\xb5\x18\x01\x12>\n" +
+	"\aWriteQC\x12\x15.storage.WriteRequest\x1a\x16.storage.WriteResponse\"\x04\xa0\xb5\x18\x01\x12E\n" +
+	"\x0eWriteMulticast\x12\x15.storage.WriteRequest\x1a\x16.google.protobuf.Empty\"\x04\x98\xb5\x18\x01B5Z.github.com/relab/gorums/examples/storage/proto\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_storage_proto_storage_proto_msgTypes = make([]protoimpl.MessageInfo, 4)
-var file_storage_proto_storage_proto_goTypes = []interface{}{
+var file_storage_proto_storage_proto_goTypes = []any{
 	(*ReadRequest)(nil),           // 0: storage.ReadRequest
 	(*ReadResponse)(nil),          // 1: storage.ReadResponse
 	(*WriteRequest)(nil),          // 2: storage.WriteRequest
@@ -343,61 +387,11 @@ func file_storage_proto_storage_proto_init() {
 	if File_storage_proto_storage_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_storage_proto_storage_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ReadRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_storage_proto_storage_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ReadResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_storage_proto_storage_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*WriteRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_storage_proto_storage_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*WriteResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_storage_proto_storage_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_storage_proto_storage_proto_rawDesc), len(file_storage_proto_storage_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   4,
 			NumExtensions: 0,
@@ -408,7 +402,6 @@ func file_storage_proto_storage_proto_init() {
 		MessageInfos:      file_storage_proto_storage_proto_msgTypes,
 	}.Build()
 	File_storage_proto_storage_proto = out.File
-	file_storage_proto_storage_proto_rawDesc = nil
 	file_storage_proto_storage_proto_goTypes = nil
 	file_storage_proto_storage_proto_depIdxs = nil
 }
