@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+ee38614a
+// 	protoc            v5.29.3
 // source: config/config.proto
 
 package config
@@ -16,9 +16,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -37,7 +37,7 @@ type Configuration struct {
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
 func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -120,7 +120,7 @@ func (m *Manager) NewConfiguration(opts ...gorums.ConfigOption) (c *Configuratio
 		}
 	}
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && c.qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -148,6 +148,14 @@ func (m *Manager) Nodes() []*Node {
 type Node struct {
 	*gorums.RawNode
 }
+
+// ConfigTestClient is the client interface for the ConfigTest service.
+type ConfigTestClient interface {
+	Config(ctx context.Context, in *Request) (resp *Response, err error)
+}
+
+// enforce interface compliance
+var _ ConfigTestClient = (*Configuration)(nil)
 
 // QuorumSpec is the interface of quorum functions for ConfigTest.
 type QuorumSpec interface {
@@ -184,11 +192,11 @@ func (c *Configuration) Config(ctx context.Context, in *Request) (resp *Response
 }
 
 // ConfigTest is the server-side API for the ConfigTest Service
-type ConfigTest interface {
+type ConfigTestServer interface {
 	Config(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 }
 
-func RegisterConfigTestServer(srv *gorums.Server, impl ConfigTest) {
+func RegisterConfigTestServer(srv *gorums.Server, impl ConfigTestServer) {
 	srv.RegisterHandler("config.ConfigTest.Config", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*Request)
 		defer ctx.Release()
