@@ -4,7 +4,6 @@ import (
 	"context"
 	fmt "fmt"
 	"testing"
-	"time"
 
 	gorums "github.com/relab/gorums"
 	"google.golang.org/grpc"
@@ -75,7 +74,7 @@ func setup(t *testing.T, mgr *Manager, cfgSize int) (cfg *Configuration, teardow
 // method on the different configurations created below.
 func TestConfig(t *testing.T) {
 	callRPC := func(cfg *Configuration) {
-		for i := 0; i < 5; i++ {
+		for i := range 5 {
 			resp, err := cfg.Config(context.Background(), &Request{Num: uint64(i)})
 			if err != nil {
 				t.Fatal(err)
@@ -86,9 +85,7 @@ func TestConfig(t *testing.T) {
 		}
 	}
 	mgr := NewManager(
-		gorums.WithDialTimeout(100*time.Millisecond),
 		gorums.WithGrpcDialOptions(
-			grpc.WithBlock(),
 			grpc.WithTransportCredentials(insecure.NewCredentials()),
 		),
 	)
