@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+ee38614a
+// 	protoc            v5.29.3
 // source: correctable/correctable.proto
 
 package correctable
@@ -18,9 +18,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -39,7 +39,7 @@ type Configuration struct {
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
 func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -122,7 +122,7 @@ func (m *Manager) NewConfiguration(opts ...gorums.ConfigOption) (c *Configuratio
 		}
 	}
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && c.qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -150,6 +150,15 @@ func (m *Manager) Nodes() []*Node {
 type Node struct {
 	*gorums.RawNode
 }
+
+// CorrectableTestClient is the client interface for the CorrectableTest service.
+type CorrectableTestClient interface {
+	Correctable(ctx context.Context, in *CorrectableRequest) *CorrectableCorrectableResponse
+	CorrectableStream(ctx context.Context, in *CorrectableRequest) *CorrectableStreamCorrectableResponse
+}
+
+// enforce interface compliance
+var _ CorrectableTestClient = (*Configuration)(nil)
 
 // Correctable asynchronously invokes a correctable quorum call on each node
 // in configuration c and returns a CorrectableCorrectableResponse, which can be used
@@ -214,12 +223,12 @@ type QuorumSpec interface {
 }
 
 // CorrectableTest is the server-side API for the CorrectableTest Service
-type CorrectableTest interface {
+type CorrectableTestServer interface {
 	Correctable(ctx gorums.ServerCtx, request *CorrectableRequest) (response *CorrectableResponse, err error)
 	CorrectableStream(ctx gorums.ServerCtx, request *CorrectableRequest, send func(response *CorrectableResponse) error) error
 }
 
-func RegisterCorrectableTestServer(srv *gorums.Server, impl CorrectableTest) {
+func RegisterCorrectableTestServer(srv *gorums.Server, impl CorrectableTestServer) {
 	srv.RegisterHandler("correctable.CorrectableTest.Correctable", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*CorrectableRequest)
 		defer ctx.Release()
