@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
-// 	protoc        v4.25.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: dummy/dummy.proto
 
 package dummy
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -21,18 +22,16 @@ const (
 )
 
 type Empty struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Empty) Reset() {
 	*x = Empty{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_dummy_dummy_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_dummy_dummy_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Empty) String() string {
@@ -43,7 +42,7 @@ func (*Empty) ProtoMessage() {}
 
 func (x *Empty) ProtoReflect() protoreflect.Message {
 	mi := &file_dummy_dummy_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -60,32 +59,27 @@ func (*Empty) Descriptor() ([]byte, []int) {
 
 var File_dummy_dummy_proto protoreflect.FileDescriptor
 
-var file_dummy_dummy_proto_rawDesc = []byte{
-	0x0a, 0x11, 0x64, 0x75, 0x6d, 0x6d, 0x79, 0x2f, 0x64, 0x75, 0x6d, 0x6d, 0x79, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x12, 0x05, 0x64, 0x75, 0x6d, 0x6d, 0x79, 0x22, 0x07, 0x0a, 0x05, 0x45, 0x6d,
-	0x70, 0x74, 0x79, 0x32, 0x2d, 0x0a, 0x05, 0x44, 0x75, 0x6d, 0x6d, 0x79, 0x12, 0x24, 0x0a, 0x04,
-	0x54, 0x65, 0x73, 0x74, 0x12, 0x0c, 0x2e, 0x64, 0x75, 0x6d, 0x6d, 0x79, 0x2e, 0x45, 0x6d, 0x70,
-	0x74, 0x79, 0x1a, 0x0c, 0x2e, 0x64, 0x75, 0x6d, 0x6d, 0x79, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79,
-	0x22, 0x00, 0x42, 0x25, 0x5a, 0x23, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
-	0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65,
-	0x73, 0x74, 0x73, 0x2f, 0x64, 0x75, 0x6d, 0x6d, 0x79, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x33,
-}
+const file_dummy_dummy_proto_rawDesc = "" +
+	"\n" +
+	"\x11dummy/dummy.proto\x12\x05dummy\"\a\n" +
+	"\x05Empty2-\n" +
+	"\x05Dummy\x12$\n" +
+	"\x04Test\x12\f.dummy.Empty\x1a\f.dummy.Empty\"\x00B*Z#github.com/relab/gorums/tests/dummy\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var (
 	file_dummy_dummy_proto_rawDescOnce sync.Once
-	file_dummy_dummy_proto_rawDescData = file_dummy_dummy_proto_rawDesc
+	file_dummy_dummy_proto_rawDescData []byte
 )
 
 func file_dummy_dummy_proto_rawDescGZIP() []byte {
 	file_dummy_dummy_proto_rawDescOnce.Do(func() {
-		file_dummy_dummy_proto_rawDescData = protoimpl.X.CompressGZIP(file_dummy_dummy_proto_rawDescData)
+		file_dummy_dummy_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_dummy_dummy_proto_rawDesc), len(file_dummy_dummy_proto_rawDesc)))
 	})
 	return file_dummy_dummy_proto_rawDescData
 }
 
 var file_dummy_dummy_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_dummy_dummy_proto_goTypes = []interface{}{
+var file_dummy_dummy_proto_goTypes = []any{
 	(*Empty)(nil), // 0: dummy.Empty
 }
 var file_dummy_dummy_proto_depIdxs = []int32{
@@ -103,25 +97,11 @@ func file_dummy_dummy_proto_init() {
 	if File_dummy_dummy_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_dummy_dummy_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Empty); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_dummy_dummy_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_dummy_dummy_proto_rawDesc), len(file_dummy_dummy_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   1,
 			NumExtensions: 0,
@@ -132,7 +112,6 @@ func file_dummy_dummy_proto_init() {
 		MessageInfos:      file_dummy_dummy_proto_msgTypes,
 	}.Build()
 	File_dummy_dummy_proto = out.File
-	file_dummy_dummy_proto_rawDesc = nil
 	file_dummy_dummy_proto_goTypes = nil
 	file_dummy_dummy_proto_depIdxs = nil
 }
