@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+ee38614a
+// 	protoc            v5.29.3
 // source: dummy/dummy.proto
 
 package dummy
@@ -15,9 +15,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -36,7 +36,7 @@ type Configuration struct {
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
 func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -119,7 +119,7 @@ func (m *Manager) NewConfiguration(opts ...gorums.ConfigOption) (c *Configuratio
 		}
 	}
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && c.qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -148,10 +148,16 @@ type Node struct {
 	*gorums.RawNode
 }
 
-// QuorumSpec is the interface of quorum functions for Dummy.
-type QuorumSpec interface {
-	gorums.ConfigOption
+// DummyNodeClient is the single node client interface for the Dummy service.
+type DummyNodeClient interface {
+	Test(ctx context.Context, in *Empty) (resp *Empty, err error)
 }
+
+// enforce interface compliance
+var _ DummyNodeClient = (*Node)(nil)
+
+// There are no quorum calls.
+type QuorumSpec interface{}
 
 // Test is a quorum call invoked on all nodes in configuration c,
 // with the same argument in, and returns a combined result.
@@ -169,11 +175,11 @@ func (n *Node) Test(ctx context.Context, in *Empty) (resp *Empty, err error) {
 }
 
 // Dummy is the server-side API for the Dummy Service
-type Dummy interface {
+type DummyServer interface {
 	Test(ctx gorums.ServerCtx, request *Empty) (response *Empty, err error)
 }
 
-func RegisterDummyServer(srv *gorums.Server, impl Dummy) {
+func RegisterDummyServer(srv *gorums.Server, impl DummyServer) {
 	srv.RegisterHandler("dummy.Dummy.Test", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*Empty)
 		defer ctx.Release()
