@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+ee38614a
+// 	protoc            v5.29.3
 // source: metadata/metadata.proto
 
 package metadata
@@ -16,9 +16,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -37,7 +37,7 @@ type Configuration struct {
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
 func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -120,7 +120,7 @@ func (m *Manager) NewConfiguration(opts ...gorums.ConfigOption) (c *Configuratio
 		}
 	}
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && c.qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -149,10 +149,17 @@ type Node struct {
 	*gorums.RawNode
 }
 
-// QuorumSpec is the interface of quorum functions for MetadataTest.
-type QuorumSpec interface {
-	gorums.ConfigOption
+// MetadataTestNodeClient is the single node client interface for the MetadataTest service.
+type MetadataTestNodeClient interface {
+	IDFromMD(ctx context.Context, in *emptypb.Empty) (resp *NodeID, err error)
+	WhatIP(ctx context.Context, in *emptypb.Empty) (resp *IPAddr, err error)
 }
+
+// enforce interface compliance
+var _ MetadataTestNodeClient = (*Node)(nil)
+
+// There are no quorum calls.
+type QuorumSpec interface{}
 
 // IDFromMD returns the 'id' field from the metadata.
 func (n *Node) IDFromMD(ctx context.Context, in *emptypb.Empty) (resp *NodeID, err error) {
@@ -183,12 +190,12 @@ func (n *Node) WhatIP(ctx context.Context, in *emptypb.Empty) (resp *IPAddr, err
 }
 
 // MetadataTest is the server-side API for the MetadataTest Service
-type MetadataTest interface {
+type MetadataTestServer interface {
 	IDFromMD(ctx gorums.ServerCtx, request *emptypb.Empty) (response *NodeID, err error)
 	WhatIP(ctx gorums.ServerCtx, request *emptypb.Empty) (response *IPAddr, err error)
 }
 
-func RegisterMetadataTestServer(srv *gorums.Server, impl MetadataTest) {
+func RegisterMetadataTestServer(srv *gorums.Server, impl MetadataTestServer) {
 	srv.RegisterHandler("metadata.MetadataTest.IDFromMD", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*emptypb.Empty)
 		defer ctx.Release()
