@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
-// 	protoc        v4.25.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: oneway/oneway.proto
 
 package oneway
@@ -12,6 +12,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -22,20 +23,17 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Num           uint64                 `protobuf:"varint,1,opt,name=Num" json:"Num,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Num uint64 `protobuf:"varint,1,opt,name=Num,proto3" json:"Num,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Request) Reset() {
 	*x = Request{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_oneway_oneway_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_oneway_oneway_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Request) String() string {
@@ -46,7 +44,7 @@ func (*Request) ProtoMessage() {}
 
 func (x *Request) ProtoReflect() protoreflect.Message {
 	mi := &file_oneway_oneway_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -69,18 +67,16 @@ func (x *Request) GetNum() uint64 {
 }
 
 type Empty struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Empty) Reset() {
 	*x = Empty{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_oneway_oneway_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_oneway_oneway_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Empty) String() string {
@@ -91,7 +87,7 @@ func (*Empty) ProtoMessage() {}
 
 func (x *Empty) ProtoReflect() protoreflect.Message {
 	mi := &file_oneway_oneway_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -108,43 +104,32 @@ func (*Empty) Descriptor() ([]byte, []int) {
 
 var File_oneway_oneway_proto protoreflect.FileDescriptor
 
-var file_oneway_oneway_proto_rawDesc = []byte{
-	0x0a, 0x13, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x2f, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x06, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x1a, 0x0c, 0x67,
-	0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x1b, 0x0a, 0x07, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x4e, 0x75, 0x6d, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x04, 0x52, 0x03, 0x4e, 0x75, 0x6d, 0x22, 0x07, 0x0a, 0x05, 0x45, 0x6d, 0x70, 0x74,
-	0x79, 0x32, 0xae, 0x01, 0x0a, 0x0a, 0x4f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x54, 0x65, 0x73, 0x74,
-	0x12, 0x2f, 0x0a, 0x07, 0x55, 0x6e, 0x69, 0x63, 0x61, 0x73, 0x74, 0x12, 0x0f, 0x2e, 0x6f, 0x6e,
-	0x65, 0x77, 0x61, 0x79, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0d, 0x2e, 0x6f,
-	0x6e, 0x65, 0x77, 0x61, 0x79, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x04, 0x90, 0xb5, 0x18,
-	0x01, 0x12, 0x31, 0x0a, 0x09, 0x4d, 0x75, 0x6c, 0x74, 0x69, 0x63, 0x61, 0x73, 0x74, 0x12, 0x0f,
-	0x2e, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
-	0x0d, 0x2e, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x04,
-	0x98, 0xb5, 0x18, 0x01, 0x12, 0x3c, 0x0a, 0x10, 0x4d, 0x75, 0x6c, 0x74, 0x69, 0x63, 0x61, 0x73,
-	0x74, 0x50, 0x65, 0x72, 0x4e, 0x6f, 0x64, 0x65, 0x12, 0x0f, 0x2e, 0x6f, 0x6e, 0x65, 0x77, 0x61,
-	0x79, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0d, 0x2e, 0x6f, 0x6e, 0x65, 0x77,
-	0x61, 0x79, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x08, 0x98, 0xb5, 0x18, 0x01, 0xa0, 0xb6,
-	0x18, 0x01, 0x42, 0x26, 0x5a, 0x24, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
-	0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65,
-	0x73, 0x74, 0x73, 0x2f, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x33,
-}
+const file_oneway_oneway_proto_rawDesc = "" +
+	"\n" +
+	"\x13oneway/oneway.proto\x12\x06oneway\x1a\fgorums.proto\"\x1b\n" +
+	"\aRequest\x12\x10\n" +
+	"\x03Num\x18\x01 \x01(\x04R\x03Num\"\a\n" +
+	"\x05Empty2\xae\x01\n" +
+	"\n" +
+	"OnewayTest\x12/\n" +
+	"\aUnicast\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\x04\x90\xb5\x18\x01\x121\n" +
+	"\tMulticast\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\x04\x98\xb5\x18\x01\x12<\n" +
+	"\x10MulticastPerNode\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\b\x98\xb5\x18\x01\xa0\xb6\x18\x01B+Z$github.com/relab/gorums/tests/oneway\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var (
 	file_oneway_oneway_proto_rawDescOnce sync.Once
-	file_oneway_oneway_proto_rawDescData = file_oneway_oneway_proto_rawDesc
+	file_oneway_oneway_proto_rawDescData []byte
 )
 
 func file_oneway_oneway_proto_rawDescGZIP() []byte {
 	file_oneway_oneway_proto_rawDescOnce.Do(func() {
-		file_oneway_oneway_proto_rawDescData = protoimpl.X.CompressGZIP(file_oneway_oneway_proto_rawDescData)
+		file_oneway_oneway_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_oneway_oneway_proto_rawDesc), len(file_oneway_oneway_proto_rawDesc)))
 	})
 	return file_oneway_oneway_proto_rawDescData
 }
 
 var file_oneway_oneway_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_oneway_oneway_proto_goTypes = []interface{}{
+var file_oneway_oneway_proto_goTypes = []any{
 	(*Request)(nil), // 0: oneway.Request
 	(*Empty)(nil),   // 1: oneway.Empty
 }
@@ -167,37 +152,11 @@ func file_oneway_oneway_proto_init() {
 	if File_oneway_oneway_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_oneway_oneway_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Request); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_oneway_oneway_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Empty); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_oneway_oneway_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_oneway_oneway_proto_rawDesc), len(file_oneway_oneway_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -208,7 +167,6 @@ func file_oneway_oneway_proto_init() {
 		MessageInfos:      file_oneway_oneway_proto_msgTypes,
 	}.Build()
 	File_oneway_oneway_proto = out.File
-	file_oneway_oneway_proto_rawDesc = nil
 	file_oneway_oneway_proto_goTypes = nil
 	file_oneway_oneway_proto_depIdxs = nil
 }
