@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+ee38614a
+// 	protoc            v5.29.3
 // source: oneway/oneway.proto
 
 package oneway
@@ -16,9 +16,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -37,7 +37,7 @@ type Configuration struct {
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
 func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -120,7 +120,7 @@ func (m *Manager) NewConfiguration(opts ...gorums.ConfigOption) (c *Configuratio
 		}
 	}
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && c.qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -148,6 +148,23 @@ func (m *Manager) Nodes() []*Node {
 type Node struct {
 	*gorums.RawNode
 }
+
+// OnewayTestClient is the client interface for the OnewayTest service.
+type OnewayTestClient interface {
+	Multicast(ctx context.Context, in *Request, opts ...gorums.CallOption)
+	MulticastPerNode(ctx context.Context, in *Request, f func(*Request, uint32) *Request, opts ...gorums.CallOption)
+}
+
+// enforce interface compliance
+var _ OnewayTestClient = (*Configuration)(nil)
+
+// OnewayTestNodeClient is the single node client interface for the OnewayTest service.
+type OnewayTestNodeClient interface {
+	Unicast(ctx context.Context, in *Request, opts ...gorums.CallOption)
+}
+
+// enforce interface compliance
+var _ OnewayTestNodeClient = (*Node)(nil)
 
 // Multicast is a quorum call invoked on all nodes in configuration c,
 // with the same argument in, and returns a combined result.
@@ -178,19 +195,17 @@ func (c *Configuration) MulticastPerNode(ctx context.Context, in *Request, f fun
 	c.RawConfiguration.Multicast(ctx, cd, opts...)
 }
 
-// QuorumSpec is the interface of quorum functions for OnewayTest.
-type QuorumSpec interface {
-	gorums.ConfigOption
-}
+// There are no quorum calls.
+type QuorumSpec interface{}
 
 // OnewayTest is the server-side API for the OnewayTest Service
-type OnewayTest interface {
+type OnewayTestServer interface {
 	Unicast(ctx gorums.ServerCtx, request *Request)
 	Multicast(ctx gorums.ServerCtx, request *Request)
 	MulticastPerNode(ctx gorums.ServerCtx, request *Request)
 }
 
-func RegisterOnewayTestServer(srv *gorums.Server, impl OnewayTest) {
+func RegisterOnewayTestServer(srv *gorums.Server, impl OnewayTestServer) {
 	srv.RegisterHandler("oneway.OnewayTest.Unicast", func(ctx gorums.ServerCtx, in *gorums.Message, _ chan<- *gorums.Message) {
 		req := in.Message.(*Request)
 		defer ctx.Release()
