@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+ee38614a
+// 	protoc            v5.29.3
 // source: ordering/order.proto
 
 package ordering
@@ -16,9 +16,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -37,7 +37,7 @@ type Configuration struct {
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
 func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -120,7 +120,7 @@ func (m *Manager) NewConfiguration(opts ...gorums.ConfigOption) (c *Configuratio
 		}
 	}
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && c.qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -168,6 +168,23 @@ func (c *Configuration) QCAsync(ctx context.Context, in *Request) *AsyncResponse
 	fut := c.RawConfiguration.AsyncCall(ctx, cd)
 	return &AsyncResponse{fut}
 }
+
+// GorumsTestClient is the client interface for the GorumsTest service.
+type GorumsTestClient interface {
+	QC(ctx context.Context, in *Request) (resp *Response, err error)
+	QCAsync(ctx context.Context, in *Request) *AsyncResponse
+}
+
+// enforce interface compliance
+var _ GorumsTestClient = (*Configuration)(nil)
+
+// GorumsTestNodeClient is the single node client interface for the GorumsTest service.
+type GorumsTestNodeClient interface {
+	UnaryRPC(ctx context.Context, in *Request) (resp *Response, err error)
+}
+
+// enforce interface compliance
+var _ GorumsTestNodeClient = (*Node)(nil)
 
 // QuorumSpec is the interface of quorum functions for GorumsTest.
 type QuorumSpec interface {
@@ -226,13 +243,13 @@ func (n *Node) UnaryRPC(ctx context.Context, in *Request) (resp *Response, err e
 }
 
 // GorumsTest is the server-side API for the GorumsTest Service
-type GorumsTest interface {
+type GorumsTestServer interface {
 	QC(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 	QCAsync(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 	UnaryRPC(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 }
 
-func RegisterGorumsTestServer(srv *gorums.Server, impl GorumsTest) {
+func RegisterGorumsTestServer(srv *gorums.Server, impl GorumsTestServer) {
 	srv.RegisterHandler("ordering.GorumsTest.QC", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*Request)
 		defer ctx.Release()
