@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+ee38614a
+// 	protoc            v5.29.3
 // source: qf/qf.proto
 
 package qf
@@ -16,9 +16,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -37,7 +37,7 @@ type Configuration struct {
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
 func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -120,7 +120,7 @@ func (m *Manager) NewConfiguration(opts ...gorums.ConfigOption) (c *Configuratio
 		}
 	}
 	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
+	var test any = struct{}{}
 	if _, empty := test.(QuorumSpec); !empty && c.qspec == nil {
 		return nil, fmt.Errorf("config: missing required QuorumSpec")
 	}
@@ -148,6 +148,15 @@ func (m *Manager) Nodes() []*Node {
 type Node struct {
 	*gorums.RawNode
 }
+
+// QuorumFunctionClient is the client interface for the QuorumFunction service.
+type QuorumFunctionClient interface {
+	UseReq(ctx context.Context, in *Request) (resp *Response, err error)
+	IgnoreReq(ctx context.Context, in *Request) (resp *Response, err error)
+}
+
+// enforce interface compliance
+var _ QuorumFunctionClient = (*Configuration)(nil)
 
 // QuorumSpec is the interface of quorum functions for QuorumFunction.
 type QuorumSpec interface {
@@ -213,12 +222,12 @@ func (c *Configuration) IgnoreReq(ctx context.Context, in *Request) (resp *Respo
 }
 
 // QuorumFunction is the server-side API for the QuorumFunction Service
-type QuorumFunction interface {
+type QuorumFunctionServer interface {
 	UseReq(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 	IgnoreReq(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 }
 
-func RegisterQuorumFunctionServer(srv *gorums.Server, impl QuorumFunction) {
+func RegisterQuorumFunctionServer(srv *gorums.Server, impl QuorumFunctionServer) {
 	srv.RegisterHandler("qf.QuorumFunction.UseReq", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*Request)
 		defer ctx.Release()
