@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"testing"
-	"time"
 
 	"github.com/relab/gorums"
 	"google.golang.org/grpc"
@@ -101,9 +100,9 @@ func BenchmarkQF(b *testing.B) {
 		request := &Request{Value: 1}
 
 		b.Run(fmt.Sprintf("UseReq_%d", n), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
+			for b.Loop() {
 				replies := make(map[uint32]*Response, mapSize)
-				for j := 0; j < n; j++ {
+				for j := range n {
 					replies[uint32(j)] = &Response{Result: request.Value}
 					resp, q := qspec.UseReqQF(request, replies)
 					if q {
@@ -113,9 +112,9 @@ func BenchmarkQF(b *testing.B) {
 			}
 		})
 		b.Run(fmt.Sprintf("IgnoreReq_%d", n), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
+			for b.Loop() {
 				replies := make(map[uint32]*Response, mapSize)
-				for j := 0; j < n; j++ {
+				for j := range n {
 					replies[uint32(j)] = &Response{Result: request.Value}
 					resp, q := qspec.IgnoreReqQF(request, replies)
 					if q {
@@ -125,9 +124,9 @@ func BenchmarkQF(b *testing.B) {
 			}
 		})
 		b.Run(fmt.Sprintf("WithoutReq_%d", n), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
+			for b.Loop() {
 				replies := make(map[uint32]*Response, mapSize)
-				for j := 0; j < n; j++ {
+				for j := range n {
 					replies[uint32(j)] = &Response{Result: request.Value}
 					resp, q := qspec.WithoutReqQF(replies)
 					if q {
@@ -140,9 +139,9 @@ func BenchmarkQF(b *testing.B) {
 		// Slice versions
 
 		b.Run(fmt.Sprintf("SliceUseReq_%d", n), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
+			for b.Loop() {
 				replies := make([]*Response, n)
-				for j := 0; j < n; j++ {
+				for j := range n {
 					replies[uint32(j)] = &Response{Result: request.Value}
 					resp, q := qspec.SliceUseReqQF(request, replies)
 					if q {
@@ -152,9 +151,9 @@ func BenchmarkQF(b *testing.B) {
 			}
 		})
 		b.Run(fmt.Sprintf("SliceIgnoreReq_%d", n), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
+			for b.Loop() {
 				replies := make([]*Response, n)
-				for j := 0; j < n; j++ {
+				for j := range n {
 					replies[uint32(j)] = &Response{Result: request.Value}
 					resp, q := qspec.SliceIgnoreReqQF(request, replies)
 					if q {
@@ -164,9 +163,9 @@ func BenchmarkQF(b *testing.B) {
 			}
 		})
 		b.Run(fmt.Sprintf("SliceWithoutReq_%d", n), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
+			for b.Loop() {
 				replies := make([]*Response, n)
-				for j := 0; j < n; j++ {
+				for j := range n {
 					replies[uint32(j)] = &Response{Result: request.Value}
 					resp, q := qspec.SliceWithoutReqQF(replies)
 					if q {
@@ -196,7 +195,6 @@ func BenchmarkFullStackQF(b *testing.B) {
 			return srv
 		})
 		mgr := NewManager(
-			gorums.WithDialTimeout(10*time.Second),
 			gorums.WithGrpcDialOptions(
 				grpc.WithTransportCredentials(insecure.NewCredentials()),
 			),
@@ -213,7 +211,7 @@ func BenchmarkFullStackQF(b *testing.B) {
 		b.ResetTimer()
 
 		b.Run(fmt.Sprintf("UseReq_%d", n), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
+			for b.Loop() {
 				resp, err := c.UseReq(context.Background(), &Request{Value: int64(requestValue)})
 				if err != nil {
 					b.Fatalf("UseReq error: %v", err)
@@ -222,7 +220,7 @@ func BenchmarkFullStackQF(b *testing.B) {
 			}
 		})
 		b.Run(fmt.Sprintf("IgnoreReq_%d", n), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
+			for b.Loop() {
 				resp, err := c.IgnoreReq(context.Background(), &Request{Value: int64(requestValue)})
 				if err != nil {
 					b.Fatalf("IgnoreReq error: %v", err)
