@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
-// 	protoc        v4.25.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: tls/tls.proto
 
 package tls
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -21,18 +22,16 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Request) Reset() {
 	*x = Request{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_tls_tls_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_tls_tls_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Request) String() string {
@@ -43,7 +42,7 @@ func (*Request) ProtoMessage() {}
 
 func (x *Request) ProtoReflect() protoreflect.Message {
 	mi := &file_tls_tls_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -59,20 +58,17 @@ func (*Request) Descriptor() ([]byte, []int) {
 }
 
 type Response struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	OK            bool                   `protobuf:"varint,1,opt,name=OK" json:"OK,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	OK bool `protobuf:"varint,1,opt,name=OK,proto3" json:"OK,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Response) Reset() {
 	*x = Response{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_tls_tls_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_tls_tls_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Response) String() string {
@@ -83,7 +79,7 @@ func (*Response) ProtoMessage() {}
 
 func (x *Response) ProtoReflect() protoreflect.Message {
 	mi := &file_tls_tls_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -107,33 +103,29 @@ func (x *Response) GetOK() bool {
 
 var File_tls_tls_proto protoreflect.FileDescriptor
 
-var file_tls_tls_proto_rawDesc = []byte{
-	0x0a, 0x0d, 0x74, 0x6c, 0x73, 0x2f, 0x74, 0x6c, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
-	0x03, 0x74, 0x6c, 0x73, 0x22, 0x09, 0x0a, 0x07, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22,
-	0x1a, 0x0a, 0x08, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x0e, 0x0a, 0x02, 0x4f,
-	0x4b, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x02, 0x4f, 0x4b, 0x32, 0x2f, 0x0a, 0x03, 0x54,
-	0x4c, 0x53, 0x12, 0x28, 0x0a, 0x07, 0x54, 0x65, 0x73, 0x74, 0x54, 0x4c, 0x53, 0x12, 0x0c, 0x2e,
-	0x74, 0x6c, 0x73, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0d, 0x2e, 0x74, 0x6c,
-	0x73, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x42, 0x23, 0x5a, 0x21,
-	0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62,
-	0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2f, 0x74, 0x6c,
-	0x73, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_tls_tls_proto_rawDesc = "" +
+	"\n" +
+	"\rtls/tls.proto\x12\x03tls\"\t\n" +
+	"\aRequest\"\x1a\n" +
+	"\bResponse\x12\x0e\n" +
+	"\x02OK\x18\x01 \x01(\bR\x02OK2/\n" +
+	"\x03TLS\x12(\n" +
+	"\aTestTLS\x12\f.tls.Request\x1a\r.tls.Response\"\x00B(Z!github.com/relab/gorums/tests/tls\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var (
 	file_tls_tls_proto_rawDescOnce sync.Once
-	file_tls_tls_proto_rawDescData = file_tls_tls_proto_rawDesc
+	file_tls_tls_proto_rawDescData []byte
 )
 
 func file_tls_tls_proto_rawDescGZIP() []byte {
 	file_tls_tls_proto_rawDescOnce.Do(func() {
-		file_tls_tls_proto_rawDescData = protoimpl.X.CompressGZIP(file_tls_tls_proto_rawDescData)
+		file_tls_tls_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_tls_tls_proto_rawDesc), len(file_tls_tls_proto_rawDesc)))
 	})
 	return file_tls_tls_proto_rawDescData
 }
 
 var file_tls_tls_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_tls_tls_proto_goTypes = []interface{}{
+var file_tls_tls_proto_goTypes = []any{
 	(*Request)(nil),  // 0: tls.Request
 	(*Response)(nil), // 1: tls.Response
 }
@@ -152,37 +144,11 @@ func file_tls_tls_proto_init() {
 	if File_tls_tls_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_tls_tls_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Request); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_tls_tls_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Response); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_tls_tls_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_tls_tls_proto_rawDesc), len(file_tls_tls_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -193,7 +159,6 @@ func file_tls_tls_proto_init() {
 		MessageInfos:      file_tls_tls_proto_msgTypes,
 	}.Build()
 	File_tls_tls_proto = out.File
-	file_tls_tls_proto_rawDesc = nil
 	file_tls_tls_proto_goTypes = nil
 	file_tls_tls_proto_depIdxs = nil
 }
