@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+903a0213
+// 	protoc            v5.29.3
 // source: benchmark/benchmark.proto
 
 package benchmark
@@ -17,9 +17,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -169,6 +169,21 @@ func (c *Configuration) AsyncQuorumCall(ctx context.Context, in *Echo) *AsyncEch
 	fut := c.RawConfiguration.AsyncCall(ctx, cd)
 	return &AsyncEcho{fut}
 }
+
+// BenchmarkClient is the client interface for the Benchmark service.
+type BenchmarkClient interface {
+	StartServerBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error)
+	StopServerBenchmark(ctx context.Context, in *StopRequest) (resp *Result, err error)
+	StartBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error)
+	StopBenchmark(ctx context.Context, in *StopRequest) (resp *MemoryStatList, err error)
+	QuorumCall(ctx context.Context, in *Echo) (resp *Echo, err error)
+	AsyncQuorumCall(ctx context.Context, in *Echo) *AsyncEcho
+	SlowServer(ctx context.Context, in *Echo) (resp *Echo, err error)
+	Multicast(ctx context.Context, in *TimedMsg, opts ...gorums.CallOption)
+}
+
+// enforce interface compliance
+var _ BenchmarkClient = (*Configuration)(nil)
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ emptypb.Empty
@@ -370,7 +385,7 @@ func (c *Configuration) SlowServer(ctx context.Context, in *Echo) (resp *Echo, e
 }
 
 // Benchmark is the server-side API for the Benchmark Service
-type Benchmark interface {
+type BenchmarkServer interface {
 	StartServerBenchmark(ctx gorums.ServerCtx, request *StartRequest) (response *StartResponse, err error)
 	StopServerBenchmark(ctx gorums.ServerCtx, request *StopRequest) (response *Result, err error)
 	StartBenchmark(ctx gorums.ServerCtx, request *StartRequest) (response *StartResponse, err error)
@@ -381,7 +396,7 @@ type Benchmark interface {
 	Multicast(ctx gorums.ServerCtx, request *TimedMsg)
 }
 
-func RegisterBenchmarkServer(srv *gorums.Server, impl Benchmark) {
+func RegisterBenchmarkServer(srv *gorums.Server, impl BenchmarkServer) {
 	srv.RegisterHandler("benchmark.Benchmark.StartServerBenchmark", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*StartRequest)
 		defer ctx.Release()
