@@ -1,23 +1,23 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v3.12.4
+// 	protoc-gen-gorums v0.9.0-devel+d62f9a0e
+// 	protoc            v5.29.3
 // source: zorums.proto
 
 package dev
 
 import (
 	context "context"
-	empty "github.com/golang/protobuf/ptypes/empty"
 	gorums "github.com/relab/gorums"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // QuorumCallAsync plain.
@@ -123,9 +123,9 @@ func (c *Configuration) QuorumCallAsyncEmpty(ctx context.Context, in *Request) *
 		Method:  "dev.ZorumsService.QuorumCallAsyncEmpty",
 	}
 	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*empty.Empty, len(replies))
+		r := make(map[uint32]*emptypb.Empty, len(replies))
 		for k, v := range replies {
-			r[k] = v.(*empty.Empty)
+			r[k] = v.(*emptypb.Empty)
 		}
 		return c.qspec.QuorumCallAsyncEmptyQF(req.(*Request), r)
 	}
@@ -136,7 +136,7 @@ func (c *Configuration) QuorumCallAsyncEmpty(ctx context.Context, in *Request) *
 
 // QuorumCallAsyncEmpty2 for testing imported message type; with same return
 // type as QuorumCallAsync: Response.
-func (c *Configuration) QuorumCallAsyncEmpty2(ctx context.Context, in *empty.Empty) *AsyncResponse {
+func (c *Configuration) QuorumCallAsyncEmpty2(ctx context.Context, in *emptypb.Empty) *AsyncResponse {
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.QuorumCallAsyncEmpty2",
@@ -146,7 +146,7 @@ func (c *Configuration) QuorumCallAsyncEmpty2(ctx context.Context, in *empty.Emp
 		for k, v := range replies {
 			r[k] = v.(*Response)
 		}
-		return c.qspec.QuorumCallAsyncEmpty2QF(req.(*empty.Empty), r)
+		return c.qspec.QuorumCallAsyncEmpty2QF(req.(*emptypb.Empty), r)
 	}
 
 	fut := c.RawConfiguration.AsyncCall(ctx, cd)
