@@ -1,23 +1,23 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v3.12.4
+// 	protoc-gen-gorums v0.9.0-devel+d62f9a0e
+// 	protoc            v5.29.3
 // source: zorums.proto
 
 package dev
 
 import (
 	context "context"
-	empty "github.com/golang/protobuf/ptypes/empty"
 	gorums "github.com/relab/gorums"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // Correctable plain.
@@ -110,9 +110,9 @@ func (c *Configuration) CorrectableEmpty(ctx context.Context, in *Request) *Corr
 		ServerStream: false,
 	}
 	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, int, bool) {
-		r := make(map[uint32]*empty.Empty, len(replies))
+		r := make(map[uint32]*emptypb.Empty, len(replies))
 		for k, v := range replies {
-			r[k] = v.(*empty.Empty)
+			r[k] = v.(*emptypb.Empty)
 		}
 		return c.qspec.CorrectableEmptyQF(req.(*Request), r)
 	}
@@ -123,7 +123,7 @@ func (c *Configuration) CorrectableEmpty(ctx context.Context, in *Request) *Corr
 
 // CorrectableEmpty2 for testing imported message type; with same return
 // type as Correctable: Response.
-func (c *Configuration) CorrectableEmpty2(ctx context.Context, in *empty.Empty) *CorrectableResponse {
+func (c *Configuration) CorrectableEmpty2(ctx context.Context, in *emptypb.Empty) *CorrectableResponse {
 	cd := gorums.CorrectableCallData{
 		Message:      in,
 		Method:       "dev.ZorumsService.CorrectableEmpty2",
@@ -134,7 +134,7 @@ func (c *Configuration) CorrectableEmpty2(ctx context.Context, in *empty.Empty) 
 		for k, v := range replies {
 			r[k] = v.(*Response)
 		}
-		return c.qspec.CorrectableEmpty2QF(req.(*empty.Empty), r)
+		return c.qspec.CorrectableEmpty2QF(req.(*emptypb.Empty), r)
 	}
 
 	corr := c.RawConfiguration.CorrectableCall(ctx, cd)
@@ -231,9 +231,9 @@ func (c *Configuration) CorrectableStreamEmpty(ctx context.Context, in *Request)
 		ServerStream: true,
 	}
 	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, int, bool) {
-		r := make(map[uint32]*empty.Empty, len(replies))
+		r := make(map[uint32]*emptypb.Empty, len(replies))
 		for k, v := range replies {
-			r[k] = v.(*empty.Empty)
+			r[k] = v.(*emptypb.Empty)
 		}
 		return c.qspec.CorrectableStreamEmptyQF(req.(*Request), r)
 	}
@@ -244,7 +244,7 @@ func (c *Configuration) CorrectableStreamEmpty(ctx context.Context, in *Request)
 
 // CorrectableEmpty2 for testing imported message type; with same return
 // type as Correctable: Response.
-func (c *Configuration) CorrectableStreamEmpty2(ctx context.Context, in *empty.Empty) *CorrectableStreamResponse {
+func (c *Configuration) CorrectableStreamEmpty2(ctx context.Context, in *emptypb.Empty) *CorrectableStreamResponse {
 	cd := gorums.CorrectableCallData{
 		Message:      in,
 		Method:       "dev.ZorumsService.CorrectableStreamEmpty2",
@@ -255,7 +255,7 @@ func (c *Configuration) CorrectableStreamEmpty2(ctx context.Context, in *empty.E
 		for k, v := range replies {
 			r[k] = v.(*Response)
 		}
-		return c.qspec.CorrectableStreamEmpty2QF(req.(*empty.Empty), r)
+		return c.qspec.CorrectableStreamEmpty2QF(req.(*emptypb.Empty), r)
 	}
 
 	corr := c.RawConfiguration.CorrectableCall(ctx, cd)
