@@ -1,21 +1,21 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v3.12.4
+// 	protoc-gen-gorums v0.9.0-devel+d62f9a0e
+// 	protoc            v5.29.3
 // source: zorums.proto
 
 package dev
 
 import (
-	empty "github.com/golang/protobuf/ptypes/empty"
 	gorums "github.com/relab/gorums"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // QuorumSpec is the interface of quorum functions for ZorumsService.
@@ -54,43 +54,22 @@ type QuorumSpec interface {
 	// quorum call method. The in parameter is the request object
 	// supplied to the QuorumCallEmpty method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *empty.Empty'.
-	QuorumCallEmptyQF(in *empty.Empty, replies map[uint32]*Response) (*Response, bool)
+	// you should implement your quorum function with '_ *emptypb.Empty'.
+	QuorumCallEmptyQF(in *emptypb.Empty, replies map[uint32]*Response) (*Response, bool)
 
 	// QuorumCallEmpty2QF is the quorum function for the QuorumCallEmpty2
 	// quorum call method. The in parameter is the request object
 	// supplied to the QuorumCallEmpty2 method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
 	// you should implement your quorum function with '_ *Request'.
-	QuorumCallEmpty2QF(in *Request, replies map[uint32]*empty.Empty) (*empty.Empty, bool)
+	QuorumCallEmpty2QF(in *Request, replies map[uint32]*emptypb.Empty) (*emptypb.Empty, bool)
 
 	// QuorumCallWithBroadcastQF is the quorum function for the QuorumCallWithBroadcast
-	// broadcast call method. The in parameter is the request object
+	// quorum call method. The in parameter is the request object
 	// supplied to the QuorumCallWithBroadcast method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
 	// you should implement your quorum function with '_ *Request'.
 	QuorumCallWithBroadcastQF(in *Request, replies map[uint32]*Response) (*Response, bool)
-
-	// BroadcastWithClientHandler1QF is the quorum function for the BroadcastWithClientHandler1
-	// broadcastcall call method. The in parameter is the request object
-	// supplied to the BroadcastWithClientHandler1 method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Request'.
-	BroadcastWithClientHandler1QF(in *Request, replies []*Response) (*Response, bool)
-
-	// BroadcastWithClientHandler2QF is the quorum function for the BroadcastWithClientHandler2
-	// broadcastcall call method. The in parameter is the request object
-	// supplied to the BroadcastWithClientHandler2 method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Request'.
-	BroadcastWithClientHandler2QF(in *Request, replies []*Response) (*Response, bool)
-
-	// BroadcastWithClientHandlerAndBroadcastOptionQF is the quorum function for the BroadcastWithClientHandlerAndBroadcastOption
-	// broadcastcall call method. The in parameter is the request object
-	// supplied to the BroadcastWithClientHandlerAndBroadcastOption method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Request'.
-	BroadcastWithClientHandlerAndBroadcastOptionQF(in *Request, replies []*Response) (*Response, bool)
 
 	// QuorumCallAsyncQF is the quorum function for the QuorumCallAsync
 	// asynchronous quorum call method. The in parameter is the request object
@@ -132,14 +111,14 @@ type QuorumSpec interface {
 	// supplied to the QuorumCallAsyncEmpty method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
 	// you should implement your quorum function with '_ *Request'.
-	QuorumCallAsyncEmptyQF(in *Request, replies map[uint32]*empty.Empty) (*empty.Empty, bool)
+	QuorumCallAsyncEmptyQF(in *Request, replies map[uint32]*emptypb.Empty) (*emptypb.Empty, bool)
 
 	// QuorumCallAsyncEmpty2QF is the quorum function for the QuorumCallAsyncEmpty2
 	// asynchronous quorum call method. The in parameter is the request object
 	// supplied to the QuorumCallAsyncEmpty2 method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *empty.Empty'.
-	QuorumCallAsyncEmpty2QF(in *empty.Empty, replies map[uint32]*Response) (*Response, bool)
+	// you should implement your quorum function with '_ *emptypb.Empty'.
+	QuorumCallAsyncEmpty2QF(in *emptypb.Empty, replies map[uint32]*Response) (*Response, bool)
 
 	// CorrectableQF is the quorum function for the Correctable
 	// correctable quorum call method. The in parameter is the request object
@@ -174,14 +153,14 @@ type QuorumSpec interface {
 	// supplied to the CorrectableEmpty method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
 	// you should implement your quorum function with '_ *Request'.
-	CorrectableEmptyQF(in *Request, replies map[uint32]*empty.Empty) (*empty.Empty, int, bool)
+	CorrectableEmptyQF(in *Request, replies map[uint32]*emptypb.Empty) (*emptypb.Empty, int, bool)
 
 	// CorrectableEmpty2QF is the quorum function for the CorrectableEmpty2
 	// correctable quorum call method. The in parameter is the request object
 	// supplied to the CorrectableEmpty2 method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *empty.Empty'.
-	CorrectableEmpty2QF(in *empty.Empty, replies map[uint32]*Response) (*Response, int, bool)
+	// you should implement your quorum function with '_ *emptypb.Empty'.
+	CorrectableEmpty2QF(in *emptypb.Empty, replies map[uint32]*Response) (*Response, int, bool)
 
 	// CorrectableStreamQF is the quorum function for the CorrectableStream
 	// correctable stream quorum call method. The in parameter is the request object
@@ -216,12 +195,12 @@ type QuorumSpec interface {
 	// supplied to the CorrectableStreamEmpty method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
 	// you should implement your quorum function with '_ *Request'.
-	CorrectableStreamEmptyQF(in *Request, replies map[uint32]*empty.Empty) (*empty.Empty, int, bool)
+	CorrectableStreamEmptyQF(in *Request, replies map[uint32]*emptypb.Empty) (*emptypb.Empty, int, bool)
 
 	// CorrectableStreamEmpty2QF is the quorum function for the CorrectableStreamEmpty2
 	// correctable stream quorum call method. The in parameter is the request object
 	// supplied to the CorrectableStreamEmpty2 method at call time, and may or may not
 	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *empty.Empty'.
-	CorrectableStreamEmpty2QF(in *empty.Empty, replies map[uint32]*Response) (*Response, int, bool)
+	// you should implement your quorum function with '_ *emptypb.Empty'.
+	CorrectableStreamEmpty2QF(in *emptypb.Empty, replies map[uint32]*Response) (*Response, int, bool)
 }
