@@ -1,23 +1,23 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v3.12.4
+// 	protoc-gen-gorums v0.9.0-devel+d62f9a0e
+// 	protoc            v5.29.3
 // source: zorums.proto
 
 package dev
 
 import (
 	context "context"
-	empty "github.com/golang/protobuf/ptypes/empty"
 	gorums "github.com/relab/gorums"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // QuorumCall plain.
@@ -111,7 +111,7 @@ func (c *Configuration) QuorumCallCombo(ctx context.Context, in *Request, f func
 }
 
 // QuorumCallEmpty for testing imported message type.
-func (c *Configuration) QuorumCallEmpty(ctx context.Context, in *empty.Empty) (resp *Response, err error) {
+func (c *Configuration) QuorumCallEmpty(ctx context.Context, in *emptypb.Empty) (resp *Response, err error) {
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.QuorumCallEmpty",
@@ -121,7 +121,7 @@ func (c *Configuration) QuorumCallEmpty(ctx context.Context, in *empty.Empty) (r
 		for k, v := range replies {
 			r[k] = v.(*Response)
 		}
-		return c.qspec.QuorumCallEmptyQF(req.(*empty.Empty), r)
+		return c.qspec.QuorumCallEmptyQF(req.(*emptypb.Empty), r)
 	}
 
 	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
@@ -132,15 +132,15 @@ func (c *Configuration) QuorumCallEmpty(ctx context.Context, in *empty.Empty) (r
 }
 
 // QuorumCallEmpty2 for testing imported message type.
-func (c *Configuration) QuorumCallEmpty2(ctx context.Context, in *Request) (resp *empty.Empty, err error) {
+func (c *Configuration) QuorumCallEmpty2(ctx context.Context, in *Request) (resp *emptypb.Empty, err error) {
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.QuorumCallEmpty2",
 	}
 	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*empty.Empty, len(replies))
+		r := make(map[uint32]*emptypb.Empty, len(replies))
 		for k, v := range replies {
-			r[k] = v.(*empty.Empty)
+			r[k] = v.(*emptypb.Empty)
 		}
 		return c.qspec.QuorumCallEmpty2QF(req.(*Request), r)
 	}
@@ -149,7 +149,7 @@ func (c *Configuration) QuorumCallEmpty2(ctx context.Context, in *Request) (resp
 	if err != nil {
 		return nil, err
 	}
-	return res.(*empty.Empty), err
+	return res.(*emptypb.Empty), err
 }
 
 // QuorumCall with broadcast option enables the server handler to broadcast.
@@ -159,9 +159,6 @@ func (c *Configuration) QuorumCallWithBroadcast(ctx context.Context, in *Request
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.QuorumCallWithBroadcast",
-
-		BroadcastID:       c.snowflake.NewBroadcastID(),
-		IsBroadcastClient: true,
 	}
 	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
 		r := make(map[uint32]*Response, len(replies))
