@@ -1,26 +1,26 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v3.12.4
+// 	protoc-gen-gorums v0.9.0-devel+d62f9a0e
+// 	protoc            v5.29.3
 // source: zorums.proto
 
 package dev
 
 import (
-	empty "github.com/golang/protobuf/ptypes/empty"
 	gorums "github.com/relab/gorums"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 type internalEmpty struct {
 	nid   uint32
-	reply *empty.Empty
+	reply *emptypb.Empty
 	err   error
 }
 
@@ -37,12 +37,12 @@ type AsyncEmpty struct {
 
 // Get returns the reply and any error associated with the called method.
 // The method blocks until a reply or error is available.
-func (f *AsyncEmpty) Get() (*empty.Empty, error) {
+func (f *AsyncEmpty) Get() (*emptypb.Empty, error) {
 	resp, err := f.Async.Get()
 	if err != nil {
 		return nil, err
 	}
-	return resp.(*empty.Empty), err
+	return resp.(*emptypb.Empty), err
 }
 
 // AsyncMyResponse is a async object for processing replies.
@@ -85,12 +85,12 @@ type CorrectableEmpty struct {
 // intermediate) reply or error is available. Level is set to LevelNotSet if no
 // reply has yet been received. The Done or Watch methods should be used to
 // ensure that a reply is available.
-func (c *CorrectableEmpty) Get() (*empty.Empty, int, error) {
+func (c *CorrectableEmpty) Get() (*emptypb.Empty, int, error) {
 	resp, level, err := c.Correctable.Get()
 	if err != nil {
 		return nil, level, err
 	}
-	return resp.(*empty.Empty), level, err
+	return resp.(*emptypb.Empty), level, err
 }
 
 // CorrectableMyResponse is a correctable object for processing replies.
@@ -139,12 +139,12 @@ type CorrectableStreamEmpty struct {
 // intermediate) reply or error is available. Level is set to LevelNotSet if no
 // reply has yet been received. The Done or Watch methods should be used to
 // ensure that a reply is available.
-func (c *CorrectableStreamEmpty) Get() (*empty.Empty, int, error) {
+func (c *CorrectableStreamEmpty) Get() (*emptypb.Empty, int, error) {
 	resp, level, err := c.Correctable.Get()
 	if err != nil {
 		return nil, level, err
 	}
-	return resp.(*empty.Empty), level, err
+	return resp.(*emptypb.Empty), level, err
 }
 
 // CorrectableStreamMyResponse is a correctable object for processing replies.
