@@ -2,7 +2,6 @@ package main
 
 import (
 	"log"
-	"time"
 
 	"github.com/relab/gorums"
 	"github.com/relab/gorums/examples/storage/proto"
@@ -10,16 +9,14 @@ import (
 	"google.golang.org/grpc/credentials/insecure"
 )
 
-func runClient(addresses []string) {
+func runClient(addresses []string) error {
 	if len(addresses) < 1 {
 		log.Fatalln("No addresses provided!")
 	}
 
 	// init gorums manager
 	mgr := proto.NewManager(
-		gorums.WithDialTimeout(1*time.Second),
 		gorums.WithGrpcDialOptions(
-			grpc.WithBlock(), // block until connections are made
 			grpc.WithTransportCredentials(insecure.NewCredentials()), // disable TLS
 		),
 	)
@@ -29,7 +26,7 @@ func runClient(addresses []string) {
 		log.Fatal(err)
 	}
 
-	Repl(mgr, cfg)
+	return Repl(mgr, cfg)
 }
 
 type qspec struct {
@@ -57,15 +54,15 @@ func (q qspec) ReadQCQF(_ *proto.ReadRequest, replies map[uint32]*proto.ReadResp
 // you should implement your quorum function with '_ *WriteRequest'.
 func (q qspec) WriteQCQF(in *proto.WriteRequest, replies map[uint32]*proto.WriteResponse) (*proto.WriteResponse, bool) {
 	// wait until at least half of the replicas have responded and have updated their value
-	if numUpdated(replies) <= q.cfgSize/2 {
-		// if all replicas have responded, there must have been another write before ours
-		// that had a newer timestamp
-		if len(replies) == q.cfgSize {
-			return &proto.WriteResponse{New: false}, true
-		}
-		return nil, false
+	if numUpdated(replies) > q.cfgSize/2 {
+		return proto.WriteResponse_builder{New: true}.Build(), true
 	}
-	return &proto.WriteResponse{New: true}, true
+	// if all replicas have responded, there must have been another write before ours
+	// that had a newer timestamp
+	if len(replies) == q.cfgSize {
+		return proto.WriteResponse_builder{New: false}.Build(), true
+	}
+	return nil, false
 }
 
 // newestValue returns the reply that had the most recent timestamp
