@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.7.0-devel
-// 	protoc            v4.25.3
+// 	protoc-gen-gorums v0.9.0-devel+903a0213
+// 	protoc            v5.29.3
 // source: storage/proto/storage.proto
 
 package proto
@@ -17,9 +17,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(7 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 7)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
@@ -150,6 +150,25 @@ type Node struct {
 	*gorums.RawNode
 }
 
+// StorageClient is the client interface for the Storage service.
+type StorageClient interface {
+	ReadQC(ctx context.Context, in *ReadRequest) (resp *ReadResponse, err error)
+	WriteQC(ctx context.Context, in *WriteRequest) (resp *WriteResponse, err error)
+	WriteMulticast(ctx context.Context, in *WriteRequest, opts ...gorums.CallOption)
+}
+
+// enforce interface compliance
+var _ StorageClient = (*Configuration)(nil)
+
+// StorageNodeClient is the single node client interface for the Storage service.
+type StorageNodeClient interface {
+	ReadRPC(ctx context.Context, in *ReadRequest) (resp *ReadResponse, err error)
+	WriteRPC(ctx context.Context, in *WriteRequest) (resp *WriteResponse, err error)
+}
+
+// enforce interface compliance
+var _ StorageNodeClient = (*Node)(nil)
+
 // Reference imports to suppress errors if they are not otherwise used.
 var _ emptypb.Empty
 
@@ -256,7 +275,7 @@ func (n *Node) WriteRPC(ctx context.Context, in *WriteRequest) (resp *WriteRespo
 }
 
 // Storage is the server-side API for the Storage Service
-type Storage interface {
+type StorageServer interface {
 	ReadRPC(ctx gorums.ServerCtx, request *ReadRequest) (response *ReadResponse, err error)
 	WriteRPC(ctx gorums.ServerCtx, request *WriteRequest) (response *WriteResponse, err error)
 	ReadQC(ctx gorums.ServerCtx, request *ReadRequest) (response *ReadResponse, err error)
@@ -264,7 +283,7 @@ type Storage interface {
 	WriteMulticast(ctx gorums.ServerCtx, request *WriteRequest)
 }
 
-func RegisterStorageServer(srv *gorums.Server, impl Storage) {
+func RegisterStorageServer(srv *gorums.Server, impl StorageServer) {
 	srv.RegisterHandler("storage.Storage.ReadRPC", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*ReadRequest)
 		defer ctx.Release()
