@@ -3,6 +3,7 @@ package version
 
 import (
 	"fmt"
+	"runtime/debug"
 	"strings"
 )
 
@@ -11,7 +12,7 @@ import (
 // Steps for cutting a new release is described in doc/release-guide.md.
 const (
 	Major      = 0
-	Minor      = 7
+	Minor      = 9
 	Patch      = 0
 	PreRelease = "devel"
 )
@@ -19,18 +20,25 @@ const (
 // String formats the version string for this module in semver format.
 //
 // Examples:
+//
 //	v1.20.1
 //	v1.21.0-rc.1
+//	v1.21.0-devel+<commit-hash>
 func String() string {
 	v := fmt.Sprintf("v%d.%d.%d", Major, Minor, Patch)
 	if PreRelease != "" {
 		v += "-" + PreRelease
 
-		// TODO: Add metadata about the commit or build hash.
-		// See https://golang.org/issue/29814
-		// See https://golang.org/issue/33533
-		// See https://github.com/golang/go/issues/37475
+		bi, ok := debug.ReadBuildInfo()
 		var metadata string
+		if ok {
+			for _, setting := range bi.Settings {
+				if setting.Key == "vcs.revision" {
+					metadata = setting.Value[0:8]
+					break
+				}
+			}
+		}
 		if strings.Contains(PreRelease, "devel") && metadata != "" {
 			v += "+" + metadata
 		}
