@@ -10,26 +10,29 @@ const (
 	// GenVersion is the runtime version required by generated .pb.go files.
 	// This is incremented when generated code relies on new functionality
 	// in the runtime.
-	GenVersion = 7
+	GenVersion = 9
 
 	// MinVersion is the minimum supported version for generated .pb.go files.
 	// This is incremented when the runtime drops support for old code.
-	MinVersion = 7
+	MinVersion = 9
 )
 
 // EnforceVersion is used by code generated by protoc-gen-gorums
 // to statically enforce minimum and maximum versions of this package.
 // A compilation failure implies either that:
-//	* the runtime package is too old and needs to be updated OR
-//	* the generated code is too old and needs to be regenerated.
+//   - the runtime package is too old and needs to be updated OR
+//   - the generated code is too old and needs to be regenerated.
 //
 // The runtime package can be upgraded by running:
+//
 //	go get github.com/relab/gorums
 //
 // The generated code can be regenerated by running:
+//
 //	protoc --gorums_out=${PROTOC_GEN_GORUMS_ARGS} ${PROTO_FILES}
 //
 // Example usage by generated code:
+//
 //	const (
 //		// Verify that this generated code is sufficiently up-to-date.
 //		_ = gorums.EnforceVersion(genVersion - gorums.MinVersion)
@@ -43,6 +46,7 @@ const (
 type EnforceVersion uint
 
 // This enforces the following invariant:
+//
 //	MinVersion ≤ GenVersion ≤ MaxVersion
 const (
 	_ = EnforceVersion(GenVersion - MinVersion)
